% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_reference.R
\name{add_to_reference}
\alias{add_to_reference}
\title{Add new clusters to an annotated reference}
\usage{
add_to_reference(
  reference_df,
  signatures_cor,
  color_attr,
  label_attr = NULL,
  new_name = NULL,
  no_label = FALSE,
  only_new_points = FALSE,
  ...
)
}
\arguments{
\item{reference_df}{The reference correlation matrix.}

\item{signatures_cor}{The signature correlation matrix as defined in
reference_signatures_correlation (but only the correlation values)}

\item{color_attr}{The labels to use for the annotation/coloring of the plot}

\item{label_attr}{In case annotation labels different than the ones of the colors are wanted, this specifies the labels to show}

\item{new_name}{The name/label that has to be given to new clusters.}

\item{no_label}{A boolean variable. If TRUE, no label is added}

\item{only_new_points}{A boolean variable that defines if only the new clusters
are labelled}

\item{...}{Additional parameters for \code{umap_plot_same_layout}}
}
\value{
The umap object and plot before and after adding the new clusters,
plus mapping quality metrics and nearest neighbor annotation
}
\description{
Given an annotated reference dataframe (reference_df),
a correlation dataframe (signatures_cor), as defined with
reference_signature_correlation, an annotation label (and possibly
label_attr), it maps the new clusters on the annotated reference
providing a UMAP object and a UMAP plot (multiple plots if focus = TRUE and
label_attr is given, with each plot that zooms-in each group
defined in color_attr, subdividing the clusters based on
label_attr). The default method of clustering the UMAP is
Louvain ('louvain'), but it could be also Walktrap ('walktrap') and
Leiden ('leiden'). A title to the UMAP is given if title is provided.
You can provide color palettes for the UMAP for both the annotation and the sub annotation
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
add_to_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
}
