% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clusters_using_markers.R
\name{annotate_clusters_using_markers}
\alias{annotate_clusters_using_markers}
\title{annotate_clusters_using_markers}
\usage{
annotate_clusters_using_markers(umap_obj, marker_list, signatures)
}
\arguments{
\item{umap_obj}{A UMAP object as given by the umap_signature_plot function or
by the add_to_annotated_reference function
(but specifying if 'New' or 'Original').
It has to be an object that contains the following chain:
\code{umap_obj$umap_out$layout}}

\item{marker_list}{A named list that contains marker genes for given cell
types}

\item{signatures}{A matrix of the signatures of the clusters in the UMAP}
}
\value{
An annotation per cluster
}
\description{
annotate_clusters_using_markers
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
umap_object <- umap_signature_plot(annotated_M)
annotate_clusters_using_markers(umap_obj = umap_object,
marker_list = list('A' = c('Gene-1', 'Gene-2'),
'B' = c('Gene-1', 'Gene-3', 'Gene-8'),
'C' = c('Gene-6', 'Gene-3')),
signatures = S)
}
