% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMapping.R
\name{annotateMapping}
\alias{annotateMapping}
\title{Add annotations to the mapped profiles}
\usage{
annotateMapping(
  net,
  new_cor,
  color_attr = "SubClass",
  col_vector = "SubClass_colors",
  n_nearest = 15,
  compute_means = FALSE
)
}
\arguments{
\item{net}{The reference network}

\item{new_cor}{The correlation matrix between the profiles and the reference
pseudobulks}

\item{color_attr}{The color attribute in \code{net}, defaults to \code{SubClass}}

\item{col_vector}{The color vector in \code{net}, defaults to \code{SubClass_colors}}

\item{n_nearest}{The number of nearest neighbors to consider for the
layout definition, mapping quality assessment and annotation. Defaults to 15}

\item{compute_means}{A boolean, if TRUE the means of the neighbors scores are
considered instead of the sum (default)}
}
\value{
And \code{S4Vectors} list containing the annotation scores for each mapped
point (\code{Annotations}), the best annotation per mapped point (\code{Best.Annotation}),
the highest scores (\code{Best.Value}), and a summary barplot (\code{Barplot})
}
\description{
Add annotations to the mapped profiles
}
