% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_edges_df.R
\name{build_edges_df}
\alias{build_edges_df}
\title{Builds the edges dataframe from an igraph network}
\usage{
build_edges_df(
  network,
  layout,
  weight_quantile = 0.75,
  weights_normalization_coef = NULL
)
}
\arguments{
\item{network}{The igraph network}

\item{layout}{A layout to define the X and Y coordinates of the edges}

\item{weight_quantile}{The quantile below which the edges will be put to NA}

\item{weights_normalization_coef}{A normalization coefficient to reduce the
width of the edges in the plots. It defaults to the number of nodes of the
network (with a maximum of 1000)}
}
\value{
The edges dataframe
}
\description{
Builds the edges dataframe from an igraph network
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
network <- compute_correlation_network(M[,seq(3,dim(M)[2])])
umap_obj <- umap_graph_clustering(M)
build_edges_df(network, layout = umap_obj$umap_out$layout)
}
