% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSimilarity.R
\name{clusterSimilarity}
\alias{clusterSimilarity}
\title{Computes the correlation of the signatures given defined groups}
\usage{
clusterSimilarity(signatures, membership)
}
\arguments{
\item{signatures}{The signatures to consider}

\item{membership}{The membership vector that defines the groups}
}
\value{
The correlation matrix and the related Heatmap
}
\description{
Computes the correlation of the signatures given defined groups
}
\examples{
#' set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
membership <- c(rep('A', 3), rep('B', 2), rep('C', 2), rep('D', 2))
clusterSimilarity(S, membership)

}
