% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_outlier_score.R
\name{compute_outlier_score}
\alias{compute_outlier_score}
\title{Compute outlier score}
\usage{
compute_outlier_score(M, n_neighbors = 15, new_clusters = NULL)
}
\arguments{
\item{M}{A numerical matrix or data frame, like those returned by
\code{reference_signatures_correlation}}

\item{n_neighbors}{The number of neighbors to use to build the SNN network}

\item{new_clusters}{If this is used to evaluate newly mapped points, the
names of the newly mapped points.}
}
\value{
A score vector for each point, the lower the score, the more likely
that point is an outlier
}
\description{
Compute outlier score
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
o <- compute_outlier_score(M)
}
