% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_snn_network.R
\name{compute_snn_network}
\alias{compute_snn_network}
\title{Computes an igraph network from the correlation matrix}
\usage{
compute_snn_network(signatures_cor, n_neighbors)
}
\arguments{
\item{signatures_cor}{The signatures correlations}

\item{n_neighbors}{The number of neighbors to consider in the sNN}
}
\value{
An igraph network
}
\description{
Computes an igraph network from the correlation matrix
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
compute_snn_network(M, n_neighbors = 10)
}
