% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialRank.R
\name{differentialRank}
\alias{differentialRank}
\title{Computes differential rank of genes in a set of signatures}
\usage{
differentialRank(
  signatures,
  membership,
  selected_groups = NULL,
  n_top_genes = 100
)
}
\arguments{
\item{signatures}{The signatures to evaluate}

\item{membership}{A membership vector of the groups to analyse. One value for
each column of the signatures}

\item{selected_groups}{A subgroup to evaluate}

\item{n_top_genes}{The number of most differentially ranked genes to return}
}
\value{
A list of the n_top_genes most differentially expressed in one group
versus all the others and in all pairwise comparisons
}
\description{
Computes differential rank of genes in a set of signatures
}
\examples{
S <- FC_signatures(matrix(runif(600000,0,30), ncol = 60))
S[,seq(1,20)] <- S[,1]
S[,seq(21,40)] <- S[,21]
S[,seq(41,60)] <- S[,41]
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
differentialRank(signatures = S, membership = c(rep('Group_1', 20),
rep('Group_2', 20),
rep('Group_3', 20)),
selected_groups = c('Group_1', 'Group_3'))
}
