% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_top_by_column.R
\name{extract_top_by_column}
\alias{extract_top_by_column}
\title{Extract the top rownames (genes) given a matrix}
\usage{
extract_top_by_column(
  mat,
  n = 5,
  decreasing = TRUE,
  vector = TRUE,
  unique = TRUE
)
}
\arguments{
\item{mat}{A numerical matrix}

\item{n}{Number of top rows}

\item{decreasing}{A boolean to tell if the top genes should be taken in
increasing or decreasing order. Defaults to TRUE.}

\item{vector}{If the top rows should be given as a single vector. Defaults
to TRUE.}

\item{unique}{Only if vector = TRUE. Boolean to tell if only the unique
top rows should be returned. Defaults to TRUE.}
}
\value{
The top rownames for each column in \code{mat}
}
\description{
Extract the top rownames (genes) given a matrix
}
\examples{
mat <- matrix(sample(seq(1,10000), 1000), ncol = 10)
colnames(mat) <- paste0('Col', seq(1, ncol(mat)))
rownames(mat) <- paste0('Row', seq(1, nrow(mat)))
extract_top_by_column(mat, n = 2)
}
