% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_pairwise.R
\name{FC_pairwise}
\alias{FC_pairwise}
\title{Computes the summed difference of a column to all the others in a matrix (signature)}
\usage{
FC_pairwise(expMat, QueryCol, RefColSet = NULL, sortedFC = FALSE)
}
\arguments{
\item{expMat}{An expression matrix}

\item{QueryCol}{A query column}

\item{RefColSet}{A subset of the matrix to consider in the difference with QueryCol}

\item{sortedFC}{A boolean variable to define if the result should be sorted (decreasing=TRUE)}
}
\value{
The summed difference between QueryCol and the other columns (the signature)
}
\description{
Computes the summed difference of a column to all the others in a matrix (signature)
}
\examples{
set.seed(123)
FC_pairwise(matrix(runif(200, 0, 10), ncol=10),
QueryCol=3,
RefColSet=c(1,4,5),
sortedFC=TRUE)
}
