% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_pairwise_xFCexclusive.R
\name{FC_pairwise_xFCexclusive}
\alias{FC_pairwise_xFCexclusive}
\title{Computes the genes that have a signature higher than a threshold}
\usage{
FC_pairwise_xFCexclusive(expMat, QueryCol, RefColSet = NULL, xFC = 1)
}
\arguments{
\item{expMat}{An expression matrix}

\item{QueryCol}{A query column}

\item{RefColSet}{A subset of the matrix to consider in the difference with QueryCol}

\item{xFC}{The difference threshold for a gene to be included}
}
\value{
The genes that pass the threshold
}
\description{
Computes the genes that have a signature higher than a threshold
}
\examples{
set.seed(123)
FC_pairwise_xFCexclusive(matrix(runif(200, 0, 10), ncol=10),
QueryCol=3,
RefColSet=c(1,4,5),
xFC=1.1)
}
