% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_signatures.R
\name{FC_signatures}
\alias{FC_signatures}
\title{Computes the signatures of an expression matrix}
\usage{
FC_signatures(expMat)
}
\arguments{
\item{expMat}{An expression matrix}
}
\value{
The signatures for each of the columns of the matrix
}
\description{
Computes the signatures of an expression matrix
}
\examples{
set.seed(123)
FC_signatures(matrix(runif(200,0,10), ncol = 10))
}
