% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_signatures_exclusiveSets.R
\name{FC_signatures_exclusiveSets}
\alias{FC_signatures_exclusiveSets}
\title{Computes the highly specific genes based on signatures}
\usage{
FC_signatures_exclusiveSets(expMat)
}
\arguments{
\item{expMat}{An expression matrix}
}
\value{
For each column in expMat, returns the genes that have a signature value higher than 1
}
\description{
Computes the highly specific genes based on signatures
}
\examples{
set.seed(123)
FC_signatures_exclusiveSets(matrix(runif(200,0,10), ncol = 10))
}
