% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dge_out.R
\name{filter_dge_out}
\alias{filter_dge_out}
\title{Filters genes based on differential expression parameters}
\usage{
filter_dge_out(
  degOut,
  cutFrac = 0.25,
  cutFC = 0.05,
  cutPadj = 0.01,
  orderOut = TRUE
)
}
\arguments{
\item{degOut}{A differential expression table as defined by run_wilcox_differential}

\item{cutFrac}{A threshold for the percentage of non-zero fature values}

\item{cutFC}{A threshold for the Fold-Change}

\item{cutPadj}{A threshold for the adjusted pvalue}

\item{orderOut}{A boolean variable to order the result based on the AUC (decreasing=TRUE)}
}
\value{
The differential expression table filtered
}
\description{
Filters genes based on differential expression parameters
}
\examples{
filter_dge_out(run_wilcox_differential(expMat = matrix(seq(1,200), ncol=10),
binGroup = c(rep(c(1,2,3),
each = 3),
4)))
}
