% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_mapping.R
\name{geneset_mapping}
\alias{geneset_mapping}
\title{GeneSet mapping}
\usage{
geneset_mapping(
  mapping_obj,
  geneset_matrix,
  show_top_n = 3,
  avoid_repetitions = TRUE,
  cluster_columns = TRUE,
  column_order = NULL
)
}
\arguments{
\item{mapping_obj}{The result of the mapping}

\item{geneset_matrix}{The matrix containing the average expression of
a set of genesets over clusters}

\item{show_top_n}{The number of genesets to show in the Heatmap with
highest signature values}

\item{avoid_repetitions}{Boolean variable
If some genesets have the same name but different
cell types, only the geneset with the first cell type will be considered}

\item{cluster_columns}{A boolean, if TRUE columns are clustered.}

\item{column_order}{The order of the columns. If NULL, no specific order is
given, the default is kept.}
}
\value{
A List with the geneset scores, the geneset scores signatures
and the Heatmap
}
\description{
GeneSet mapping
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
mapping_obj <- add_to_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
anno_matrix <- matrix(runif(100,0,10), ncol = 10)
colnames(anno_matrix) <- colnames(S)
rownames(anno_matrix) <- paste0('Annotation-', seq(1, dim(anno_matrix)[1]))
anno_matrix <- t(t(anno_matrix)/colSums(anno_matrix))
geneset_mapping(mapping_obj, anno_matrix, avoid_repetitions = FALSE)
}
