% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_subclass_vs_stage.R
\name{get_average_subclass_vs_stage}
\alias{get_average_subclass_vs_stage}
\title{Get average expression divided into subclass and signatures}
\usage{
get_average_subclass_vs_stage(net, genes)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the average expression}
}
\value{
A matrix with the average expression by subclass and stage
}
\description{
Get average expression divided into subclass and signatures
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
neuRoDev:::get_average_subclass_vs_stage(net = net, genes = rownames(m)[seq(1,3)])
}
