% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_group_average.R
\name{get_column_group_average}
\alias{get_column_group_average}
\title{Gets the average by-column of columns belonging to a certain group (profiles)}
\usage{
get_column_group_average(M, group, na.rm = TRUE)
}
\arguments{
\item{M}{A matrix}

\item{group}{A membership vector for the columns of M}

\item{na.rm}{If NAs must be removed in the calculation. Defaults to TRUE}
}
\value{
A table with a single entry for each group with the profile.
}
\description{
This function computes the column average of each column belonging to a given
group defined by the membership vector (profiles).
}
\examples{
neuRoDev:::get_column_group_average(M = matrix(seq(1,100), ncol=10),
group = c(rep(c('A','B','C'), each = 3), 'D'))
neuRoDev:::get_column_group_average(M = matrix(seq(1,100), ncol=10),
group = c(rep(c('A','B','C'), each = 3), 'D'))
}
