% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_group_detection_rate.R
\name{get_column_group_detection_rate}
\alias{get_column_group_detection_rate}
\title{Defines the portion of members of each group that have non-zero values in each row}
\usage{
get_column_group_detection_rate(M, group, doParallel = FALSE)
}
\arguments{
\item{M}{A matrix}

\item{group}{A membership vector for columns of M}

\item{doParallel}{A boolean variable to run the code in parallel}
}
\value{
The portion of members of each group that have non-zero value in each row
}
\description{
Defines the portion of members of each group that have non-zero values in each row
}
\examples{
get_column_group_detection_rate(matrix(sample(seq(0,1),
100,
replace=TRUE),
ncol=10),
group=c(rep(c(1,2,3), each = 3), 4))
}
