% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlation_values.R
\name{get_correlation_values}
\alias{get_correlation_values}
\title{Get correlation values from a correlation_df object}
\usage{
get_correlation_values(correlation_df)
}
\arguments{
\item{correlation_df}{A correlation dataframe obtained from
\code{reference_signatures_correlation}}
}
\value{
Only the correlation values as a matrix
}
\description{
Get correlation values from a correlation_df object
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
cor_df <- reference_signatures_correlation(S, refS)
get_correlation_values(cor_df)
}
