% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eMatrix.R
\name{get_eMatrix}
\alias{get_eMatrix}
\title{Get the expression enrichment matrix}
\usage{
get_eMatrix(net, genes, nRand = 100)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the expression enrichment}

\item{nRand}{The number of random sampling to use (defaults to 100)}
}
\value{
A list with z-scores and p-values matrices divided into
subclass-stage pairs
}
\description{
Get the expression enrichment matrix
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
emat <- get_eMatrix(net = net, genes = rownames(net)[seq(1,5)])
}
