% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eTrace.R
\name{get_eTrace}
\alias{get_eTrace}
\title{Get the expression enrichment Trace}
\usage{
get_eTrace(net, genes, nRand = 100)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the expression enrichment}

\item{nRand}{The number of random sampling to use (defaults to 100)}
}
\value{
An S4Vectors List with the averaged scaled logcounts of the genes (\code{obs}),
the z score of the enrichment (\code{z}), and the pvalue (\code{p})
}
\description{
Get the expression enrichment Trace
}
