% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneset_network.R
\name{get_geneset_network}
\alias{get_geneset_network}
\title{Create the gene set network}
\usage{
get_geneset_network(
  geneset_list,
  pseudo,
  net_df,
  stages_palette,
  stages = "Stages",
  resolution = 4,
  padj_threshold = 0.05,
  gsea = NULL,
  between_class_pseudo_enrichment = NULL,
  between_class_gsea = NULL,
  pseudo_de = NULL,
  max.words = 10,
  ...
)
}
\arguments{
\item{geneset_list}{The list of annotated gene sets}

\item{pseudo}{The pseudobulk in form of a SingleCellExperiment object (it
needs to have the logcounts attribute)}

\item{net_df}{The reference network dataframe (the one which contains
the correlation values with the reference signatures, obtained from
\code{reference_signatures_correlation})}

\item{stages_palette}{The palette for the visualization of the stages.}

\item{stages}{The attribute in which to find the time information (stages) in
the \code{net_df}. Defaults to 'Stages'.}

\item{resolution}{The resolution for clustering in igraph::cluster_leiden.
Defaults to 4.}

\item{padj_threshold}{The adjusted pvalue threshold to consider one
geneset to be significant. Defaults to 0.05}

\item{gsea}{If the GSEA was already performed, it can be given here and not
recomputed.}

\item{between_class_pseudo_enrichment}{If an additional pseudobulk enrichment
is performed, for example between classes, it can be given here as a single
vector for the wanted class.}

\item{between_class_gsea}{If the GSEA on the additional pseudobulk enrichment
was already performed before, it can be given here and not recomputed.}

\item{pseudo_de}{If the pseudobulk enrichment between time points (stages)
was already computed, it can be given here and not recomputed.}

\item{max.words}{The maximum number of keywords per cluster to be returned.}

\item{...}{Additional \code{fgsea} parameters}
}
\value{
A list that contains the GSEA(s), the geneset network and an
geneset dataframe, with the genesets, geneset groups, the pvalue
adjusted of the GSEA(s) (one column per GSEA), the stage in which the group
has the maximum expression, the leading edges of each geneset and the top
10 most frequent genes for each geneset group.
}
\description{
Create the gene set network
}
\examples{
geneset_list <- list('A' = c('Gene-1', 'Gene-2', 'Gene-3'),
'B' = c('Gene-1', 'Gene-2', 'Gene-4'),
'C' = c('Gene-3', 'Gene-4', 'Gene-1'))
pseudo <- SingleCellExperiment::SingleCellExperiment(
assays = list(logcounts = matrix(sample(seq(1,100),
4*10,
replace = TRUE),
nrow = 4)))
rownames(pseudo) <- paste0('Gene-', seq(1, nrow(pseudo)))
colnames(pseudo) <- paste0('Column-', seq(1, ncol(pseudo)))
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
net_df <- reference_signatures_correlation(S, refS)
net_df$Stages <- paste0('Stage.', c(1, 1, 2, 2, 2))
stages_palette <- c('Stage.1' = 'red', 'Stage.2' = 'blue')
get_geneset_network(geneset_list, pseudo, net_df)
}
