% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_largest_connected_component.R
\name{get_largest_connected_component}
\alias{get_largest_connected_component}
\title{Get largest connected component}
\usage{
get_largest_connected_component(Net)
}
\arguments{
\item{Net}{An igraph Network}
}
\value{
The largest connected component of the netwokr
}
\description{
Get largest connected component
}
\examples{
set.seed(123)
num_vertices <- 100
p <- 0.3
random_graph <- igraph::erdos.renyi.game(num_vertices, p, directed = FALSE)
get_largest_connected_component(random_graph)
}
