% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix_from_list.R
\name{get_matrix_from_list}
\alias{get_matrix_from_list}
\title{Get a matrix from a list}
\usage{
get_matrix_from_list(l)
}
\arguments{
\item{l}{A list of named numerical vectors}
}
\value{
A matrix with as columns each name of any individual named vector and as
}
\description{
Get a matrix from a list
}
\examples{
v1 <- seq(2,12)
names(v1) <- c(rep('A', 5), rep('B', 6))
v2 <- seq(0,10)
names(v2) <- c(rep('B', 5), rep('C', 6))
l <- list('X' = v1, 'Y' = v2)
neuRoDev:::get_matrix_from_list(l)
}
