% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_neighbor_knowledge.R
\name{get_nearest_neighbor_knowledge}
\alias{get_nearest_neighbor_knowledge}
\title{Get nearest neighbor knowledge}
\usage{
get_nearest_neighbor_knowledge(
  reference_df,
  new_clusters,
  umap_obj,
  color_attr,
  col_vector = NULL,
  n_nearest = NULL,
  ylim = NULL,
  title = NULL,
  to_exclude = NULL,
  compute_means = FALSE
)
}
\arguments{
\item{reference_df}{The reference correlation matrix.}

\item{new_clusters}{The new clusters names.}

\item{umap_obj}{An already computed umap object as returned by
umap_graph_clustering}

\item{color_attr}{The labels to use for the annotation}

\item{col_vector}{A named color palette vector for the annotations}

\item{n_nearest}{The number of nearest neighbors to consider. It defaults
to \code{((dim(distance_matrix)[2])-1)}, where distance_matrix is given by:
\code{umap_obj$umap_out$knn$distances} or
\code{umap_obj$umap_out$refined_network$distances}}

\item{ylim}{The Y limits of the boxplot}

\item{title}{The title to add to the plots}

\item{to_exclude}{The labels inside \code{reference_df[,color_attr]}
that the user doesn't want to consider.}

\item{compute_means}{A boolean variable. If TRUE, it computes the mean instead
of the sum of values belonging to a given group.}
}
\value{
A list of the whole distribution of the variable in the nearest
neighbors, the whole frequency and whole percentages, then also divided by
cluster, plus a Heatmap and a barplot if the label is a character, a boxplot
for the whole distribution and one per new cluster if it is numeric
}
\description{
Given an annotated reference dataframe reference_df,
a set of new clusters to analyze in new_clusters, a umap object as run
with add_to_reference, which contains the new clusters to
analyze, and a given annotation label, it returns information about the
nearest neighbors for each cluster and as a whole for that given
annotation label. If remove_low_confidence=TRUE, it doesn't consider in the
analysis low confidence neighbors, as defined with the distance_threshold
parameter. You can provide a col_vector palette for the final barplot or
boxplot.
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
res <- add_to_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
umap_obj <- res$New
new_clusters = rownames(new_M)
get_nearest_neighbor_knowledge(annotated_M,
new_clusters,
umap_obj,
color_attr = 'Best.Assignment')
}
