% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_localization.R
\name{get_network_localization}
\alias{get_network_localization}
\title{Given an expression matrix and a snn graph in the form of an adjacency matrix
it returns the network localization value for each gene}
\usage{
get_network_localization(expression_matrix, snn_graph)
}
\arguments{
\item{expression_matrix}{The expression matrix}

\item{snn_graph}{The adjacency matrix of the SNN graph (or other graphs)}
}
\value{
The network localization values for each gene
}
\description{
Given an expression matrix and a snn graph in the form of an adjacency matrix
it returns the network localization value for each gene
}
\examples{
snn_graph <- matrix(sample(seq(1, 100), 10000, replace = TRUE), ncol = 100)
colnames(snn_graph) <- paste0('Column', seq(1, 100))
rownames(snn_graph) <- paste0('Column', seq(1, 100))
expression_matrix <- matrix(sample(seq(1, 100), 100000, replace = TRUE),
ncol = 100)
colnames(expression_matrix) <- paste0('Column', seq(1, 100))
rownames(expression_matrix) <- paste0('Gene', seq(1, 1000))
get_network_localization(expression_matrix, snn_graph)
}
