% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairwise_wilcoxauc.R
\name{get_pairwise_wilcoxauc}
\alias{get_pairwise_wilcoxauc}
\title{Computes the Wilcoxon test on a group of an expression matrix against the others}
\usage{
get_pairwise_wilcoxauc(expMat, group, qG, orderOut = TRUE)
}
\arguments{
\item{expMat}{An expression matrix}

\item{group}{A membership vector for the columns of expMat}

\item{qG}{A query group}

\item{orderOut}{A boolean variable to determine if the output is ordered
based on the AUC (decreasing=TRUE)}
}
\value{
The results of the Wilcoxon test: a dataframe of the differential
expression analysis with the main
statistics, like the p-value of each gene, the Bonferroni adjusted pvalue (padj),
the log Fold-Change, the AUC, the average expression (avgExpr), the percentage of
non zero values (pct_in), and a score which is the -log10 of the adjusted
pvalue multiplied by the sign of the logFC.
}
\description{
Computes the Wilcoxon test on a group of an expression matrix against the others
}
\examples{
set.seed(123)
expMat = matrix(runif(200,0,10), ncol = 10)
rownames(expMat) <- paste0('Gene-', seq(1, nrow(expMat)))
get_pairwise_wilcoxauc(expMat = expMat,
group = c(rep(c(1,2,3), each = 3), 4), qG = 2)
}
