% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairwise_wilcoxauc_DEs.R
\name{get_pairwise_wilcoxauc_DEs}
\alias{get_pairwise_wilcoxauc_DEs}
\title{Computes the pairwise Wilcoxon test and filters the results if wanted}
\usage{
get_pairwise_wilcoxauc_DEs(expMat, group, filterDEset = TRUE)
}
\arguments{
\item{expMat}{An expression matrix}

\item{group}{A membership vector for the columns of expMat}

\item{filterDEset}{A boolean variable that indicates if the function
keep_exclusive_DEGs has to be run, keeping only exclusive genes for each group}
}
\value{
Returns the genes for each group that are exclusive
}
\description{
Computes the pairwise Wilcoxon test and filters the results if wanted
}
\examples{
set.seed(123)
get_pairwise_wilcoxauc_DEs(expMat = matrix(runif(500,0,100), ncol = 10),
c(rep(c(1,2,3), each = 3), 4))
}
