% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{get_palette}
\alias{get_palette}
\title{Get the palette}
\usage{
get_palette(net, color_attr = "SubClass")
}
\arguments{
\item{net}{The reference network}

\item{color_attr}{Either SubClass or Stages (to get the palette of the SubClass
or Stages, respectively)}
}
\value{
The palette
}
\description{
Get the palette
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_color <- subclass_palette[net$SubClass]
neuRoDev:::get_palette(net)
}
