% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pseudobulk.R
\name{get_pseudobulk}
\alias{get_pseudobulk}
\title{Get pseudobulk from an expression matrix}
\usage{
get_pseudobulk(exp_matrix, membership_vector)
}
\arguments{
\item{exp_matrix}{The expression matrix (raw counts)}

\item{membership_vector}{The membership vector on which to group the
\code{exp_matrix} (e.g., clusters)}
}
\value{
A SingleCellExperiment object with the pseudobulk counts and normalized
logcounts
}
\description{
Get pseudobulk from an expression matrix
}
\examples{
exp_matrix <- matrix(sample(seq(1,10, length.out=1000), 100000, replace = TRUE), ncol = 100)
membership_vector <- c(rep('A', 25), rep('B', 25), rep('C', 25), rep('D', 25))
pseudobulk <- get_pseudobulk(exp_matrix, membership_vector)
}
