% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pseudobulk_preferential_expression.R
\name{get_pseudobulk_preferential_expression}
\alias{get_pseudobulk_preferential_expression}
\title{Obtain the pseudobulk preferential expression}
\usage{
get_pseudobulk_preferential_expression(
  expression_matrix,
  groups,
  Datatype = "RNAseq",
  threshold = 0.1
)
}
\arguments{
\item{expression_matrix}{The expression matrix to use}

\item{groups}{The groups in which to divide the expression matrix}

\item{Datatype}{The expression_matrix type. Defaults to RNAseq.}

\item{threshold}{The threshold for the pvalue above which to put the coef
returned by limma::eBayes to 0.}
}
\value{
The pseudobulk preferential expression scores
}
\description{
Obtain the pseudobulk preferential expression
}
\examples{
expression_matrix <- matrix(sample(seq(1,100), 10000, replace = TRUE), ncol = 10)
rownames(expression_matrix) <- paste0('Gene.', seq(1, nrow(expression_matrix)))
colnames(expression_matrix) <- paste0('Column.', seq(1, ncol(expression_matrix)))
groups <- paste0('Group.', c(1,1,2,3,3,4,5,4,2,5))
get_pseudobulk_preferential_expression(expression_matrix, groups)
}
