% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specificity.R
\name{get_specificity}
\alias{get_specificity}
\title{Given an expression matrix, it returns the specificity value for each gene}
\usage{
get_specificity(expression_matrix)
}
\arguments{
\item{expression_matrix}{An expression matrix}
}
\value{
The specificity values for each gene
}
\description{
Given an expression matrix, it returns the specificity value for each gene
}
\examples{
get_specificity(matrix(sample(seq(1,100), 10000, replace = TRUE),
ncol = 100))
}
