% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subclass_score.R
\name{get_subclass_score}
\alias{get_subclass_score}
\title{Get subclass score for mapping}
\usage{
get_subclass_score(mapped_obj, reference_df, sub_idxs = NULL)
}
\arguments{
\item{mapped_obj}{The result of mapData}

\item{reference_df}{The dataframe with the annotations per reference cluster}

\item{sub_idxs}{The indexes from which to select the annotations scores in
mapped_obj$NearestNeighborsAnnotation$Annotations. If NULL (default), they are
all kept.}
}
\value{
A score per subclass considering all the mapped data points
}
\description{
Get subclass score for mapping
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
mapped_obj <- mapData(reference_df = annotated_M,
new_signatures = new_clusterS,
reference_signatures = refS,
color_attr = 'Best.Assignment')
get_subclass_score(mapped_obj, reference_df = annotated_M)
}
