% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterSignatures.R
\name{getClusterSignatures}
\alias{getClusterSignatures}
\title{Signatures Initialization}
\usage{
getClusterSignatures(
  expression_matrix,
  algorithm = "louvain",
  split = NULL,
  subsample = FALSE,
  PCs_to_use = NULL,
  resolution = 1,
  clusters_mv = NULL,
  Ntotal = 2000,
  BOGs = FALSE,
  pBOGs = FALSE,
  rEEA = FALSE,
  pBOGsig = FALSE,
  verbose = FALSE,
  condaenv = "r-reticulate"
)
}
\arguments{
\item{expression_matrix}{A single-cell RNAseq matrix or a list of matrices}

\item{algorithm}{The algorithm to use between 'louvain' and 'leiden' to
cluster cells}

\item{split}{A vector that defines how to split a expression_matrix into multiple ones}

\item{subsample}{A boolean variable that defines if each of the provided
expression_matrixs has to be proportionally subsampled to the same number of cells}

\item{PCs_to_use}{The number of Principal Components to consider for
FindNeighbors of Seurat}

\item{resolution}{The resolution to consider for FindClusters of Seurat}

\item{clusters_mv}{A vector that substitutes the clustering
(ex. celltypes if already annotated)}

\item{Ntotal}{The total number of cells to consider in the subsampling}

\item{BOGs}{A boolean variable to define if map_clusters_to_representation has
to compute also the BOGs}

\item{pBOGs}{A boolean variable to define if map_clusters_to_representation has
to compute also the pBOGs}

\item{rEEA}{A boolean variable to define if map_clusters_to_representation has
to compute also the rEEA}

\item{pBOGsig}{A boolean variable to define if map_clusters_to_representation
has to compute also the pBOGsig}

\item{verbose}{A boolean variable to define if the steps of Seurat will be
printed or not}

\item{condaenv}{The name of the conda environment in which 'leidenalg'
module is present, in case the user wants to use the Leiden algorithm}
}
\value{
The result of map_clusters_to_representation run on each expression_matrix after
splitting, subsampling, clustering if/when required
}
\description{
Signatures Initialization
}
\examples{
set.seed(123)
M <- matrix(runif(2000000,0,100), ncol=20000)
rownames(M) <- paste0('Gene-', seq(1, dim(M)[1]))
colnames(M) <- paste0('Cell-', seq(1, dim(M)[2]))
getClusterSignatures(M, resolution=2)
}
