% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrivingGenes.R
\name{getDrivingGenes}
\alias{getDrivingGenes}
\title{Get genes driving the newly mapped points}
\usage{
getDrivingGenes(
  new_umap_obj,
  newSignatures,
  n_genes = 100,
  threshold = 0.9,
  gene_set = NULL
)
}
\arguments{
\item{new_umap_obj}{The umap object coming from add_to_reference,
under \code{New$umap_obj}}

\item{newSignatures}{The signatures of the new points}

\item{n_genes}{The number of top genes for each new point to be returned}

\item{threshold}{The signatures level threshold under which genes are not
considered. It's specific for signatures of each new point.}

\item{gene_set}{A starting gene set from which to extract the top genes}
}
\value{
A list of genes for each new point mapped
}
\description{
Get genes driving the newly mapped points
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
mapped <- add_to_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
getDrivingGenes(mapped$New$umap_obj, new_clusterS)
}
