% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupDistance.R
\name{getGroupDistance}
\alias{getGroupDistance}
\title{getGroupDistance}
\usage{
getGroupDistance(annotated_reference, membership_label, umap_obj)
}
\arguments{
\item{annotated_reference}{The annotated reference, like those returned by
reference_signatures_correlation}

\item{membership_label}{A membership label that identifies a column in the
annotated reference}

\item{umap_obj}{A umap object like those returned by umap_signature_plot}
}
\value{
Three matrices and three heatmaps, one for the mean distance, one for
the max distance and one for the min distance
}
\description{
It computes the mean, max and minimum distances between clusters of each group
given a membership label, an annotated reference (like those returned after
correlating signatures) and a umap object
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
group = M$`Best.Assignment`
umap_obj = umap_signature_plot(M,
color_attr = group,
label_attr = group)
getGroupDistance(M, 'Best.Assignment', umap_obj)
}
