% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMappingConfidence.R
\name{getMappingConfidence}
\alias{getMappingConfidence}
\title{getMappingConfidence}
\usage{
getMappingConfidence(mapped_obj, signatures_cor)
}
\arguments{
\item{mapped_obj}{The full mapped object derived from umap_plot_same_layout or
add_to_reference (containing both the New and Original elements)}

\item{signatures_cor}{The signature correlation matrix as defined in
reference_signatures_correlation}
}
\value{
The confidence and mappability of the given mapping
}
\description{
getMappingConfidence
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_profiles <- matrix(runif(80,0,10), ncol = 4)
rownames(new_profiles) <- paste0('Gene-', seq(1, dim(new_profiles)[1]))
colnames(new_profiles) <- paste0('New-', seq(1, dim(new_profiles)[2]))
new_clusterS <- FC_signatures(new_profiles)
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
mapping <- add_to_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
getMappingConfidence(mapped_obj = mapping, signatures_cor = new_M)
}
