% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeaningfulGenes.R
\name{getMeaningfulGenes}
\alias{getMeaningfulGenes}
\title{Identifies the genes responsible for the mapping of new clusters onto the UMAP}
\usage{
getMeaningfulGenes(
  refSignatures,
  newSignatures,
  umap_obj,
  new_clusters,
  n_genes = 20,
  threshold = 0.99,
  n_nearest = NULL
)
}
\arguments{
\item{refSignatures}{The signatures of the clusters in the UMAP (not new ones)}

\item{newSignatures}{The signatures of the new clusters}

\item{umap_obj}{A UMAP object as given by the add_to_annotated_reference
function (but specifying 'New').
It has to be an object that contains the following chain:
\code{umap_obj$umap_obj$umap_out$layout} or \code{umap_obj$umap_out$layout}}

\item{new_clusters}{The names of the new clusters}

\item{n_genes}{The number of genes to be returned}

\item{threshold}{A threshold for the quality of the genes (hence, genes might
be less than n_genes in the end)}

\item{n_nearest}{Number of nearest neighbors to consider. It defaults to the
number of nearest neighbors used to construct the UMAP}
}
\value{
A list of lists, in which each element of the first list considers one
new cluster and it gives a list with the the impactful genes and the
confidence
}
\description{
Identifies the genes responsible for the mapping of new clusters onto the UMAP
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(100,0.1,7), ncol = 5))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
to_reference <- add_to_annotated_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
getMeaningfulGenes(umap_obj = to_reference$New,
new_clusters = rownames(new_M),
refSignatures = S,
newSignatures = new_clusterS,
threshold = 0.5)
}
