% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignatureMeans.R
\name{getSignatureMeans}
\alias{getSignatureMeans}
\title{Computes the mean rank of all genes across signatures}
\usage{
getSignatureMeans(signatures, n_genes, weights = NULL, new_cluster = NULL)
}
\arguments{
\item{signatures}{The wanted signatures}

\item{n_genes}{The number of top genes to be highlighted}

\item{weights}{Possible weights to have a weighted mean}

\item{new_cluster}{Possible indication of a new cluster whose ranks will be
averaged against the average of all other clusters (so will matter more)}
}
\value{
A SimpleList of the selected top genes and the mean rank of all genes
}
\description{
Computes the mean rank of all genes across signatures
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
getSignatureMeans(S, n_genes = 10)
}
