% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gliogenesis.sce.R
\name{gliogenesis.sce}
\alias{gliogenesis.sce}
\title{SingleCellExperiment object}
\format{
A \code{SingleCellExperiment} object with:
\describe{
\item{assays}{A list containing one matrix named "signatures" (genes x clusters),
one matrix named "logcounts" (genes x clusters)}
\item{colData}{A \code{DataFrame} with clusters-level metadata (stage, subclass,
study etc.)}
\item{metadata}{A list containing one \code{SingleCellExperiment} object with
subclass-level pseudobulks (\code{subclass_psb}), one \code{SingleCellExperiment} object
with stage-level pseudobulks (\code{stage_psb}), one correlation matrix with
cell type reference signatures (\code{correlations}), one \code{List} object with the
network information (\code{network})}
}
}
\source{
Obtained by processing the gliogenesis lineage of the single-cell
data compendium.
}
\usage{
gliogenesis.sce(directory = NULL)
}
\arguments{
\item{directory}{The directory in which to download the object or from which
to load it. If NULL (default), it downloads it in the cache of the neuRoDev
package.}
}
\description{
Function to automatically download the gliogenesis reference network along
with the expression profiles and metadata. If the object was previously
downloaded with this function, the function will also directly load the
object.
}
