% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_diffusion.R
\name{heat_diffusion}
\alias{heat_diffusion}
\title{heat_diffusion}
\usage{
heat_diffusion(graph, scores, alpha = 0.7, tol = 1e-06, max_iter = 100)
}
\arguments{
\item{graph}{The network on which to diffuse}

\item{scores}{The starting scores to diffuse}

\item{alpha}{Alpha value to give to expm function}

\item{tol}{a given tolerance used to check if \code{x} is
    computationally singular when \code{method = "hybrid_Eigen_Ward"}.}

\item{max_iter}{Maximum number of iterations}
}
\value{
The diffused scores
}
\description{
heat_diffusion
}
\examples{
g <- igraph::make_ring(10)
igraph::E(g)$weight <- runif(igraph::ecount(g), 0.5, 1)
initial_scores <- rep(0, igraph::vcount(g))
initial_scores[1] <- 1
heat_diffusion(g, initial_scores)
}
