% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_number.R
\name{is_number}
\alias{is_number}
\title{Defines if a character is a number}
\usage{
is_number(x)
}
\arguments{
\item{x}{A character}
}
\value{
TRUE if it's a number, FALSE otherwise
}
\description{
Defines if a character is a number
}
\examples{
neuRoDev:::is_number('1')
}
