% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_exclusive_DEGs.R
\name{keep_exclusive_DEGs}
\alias{keep_exclusive_DEGs}
\title{Keeps only genes exclusively differentially expressed in each group}
\usage{
keep_exclusive_DEGs(pwOutList)
}
\arguments{
\item{pwOutList}{The output of get_pairwise_wilcoxauc filtered with filter_dge_out}
}
\value{
The genes for each group that are exclusive for that group
}
\description{
Keeps only genes exclusively differentially expressed in each group
}
\examples{
group <- c(rep(c(1,2,3), each = 3), 4)
out <- lapply(unique(group), function(i) {
o <- get_pairwise_wilcoxauc(expMat = matrix(runif(500,0,100),
ncol = 10), group = group, qG = i)
pDEGs <- lapply(o, filter_dge_out, cutFrac=0.01, cutFC=0.0, cutPadj=1)
return(pDEGs)
})

keep_exclusive_DEGs(out)
}
