% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_diffusion.R
\name{label_diffusion}
\alias{label_diffusion}
\title{Label diffusion K-medoids like}
\usage{
label_diffusion(
  umap_obj,
  labels,
  max_n_iter = 1000,
  max_n_equal = 10,
  perc_equal = 0.99,
  n_nearest = NULL
)
}
\arguments{
\item{umap_obj}{The starting umap object}

\item{labels}{The starting labels vector (named with the vertex/cluster names
if you suspect that the network stored in the umap object doesn't have the
same order of vertexes as those that are in the labels vector)}

\item{max_n_iter}{Maximum number of iterations}

\item{max_n_equal}{The number of times the previous association has to be
equal (to a percentage dictated by \code{perc_equal}) to the current association
to stop the diffusion}

\item{perc_equal}{The percentage of same association between the different
iterations to be considered equal}

\item{n_nearest}{The number of nearest neighbours to consider for
classification. It defaults to the minimum between the number of labels / 100
and half of the label group with the lowest number of components.}
}
\value{
The diffused labels
}
\description{
Label diffusion K-medoids like
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
umap_obj <- umap_graph_clustering(M)
graph <- umap_obj$refined_network
igraph::V(graph)$label <- paste0('Group', rep(seq(1,2), each = 5))
label_diffusion(umap_obj = umap_obj,
labels = paste0('Group', rep(seq(1,2), each = 5)))
}
