% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelPrediction.R
\name{labelPrediction}
\alias{labelPrediction}
\title{Prediction of previous clusters given new ones}
\usage{
labelPrediction(
  umap_obj,
  new_clusters,
  col_vector = NULL,
  new_points_col = NULL,
  legend = FALSE,
  layout = NULL,
  update_reference = FALSE,
  reference = NULL,
  weight_quantile = 0.75,
  weights_normalization_coef = NULL,
  show_edges = TRUE,
  only_best_association = FALSE
)
}
\arguments{
\item{umap_obj}{An already computed umap object as returned by
umap_graph_clustering}

\item{new_clusters}{The new clusters names.}

\item{col_vector}{A named color vector palette with a color for each color_attr}

\item{new_points_col}{The color of the new clusters in the UMAP plot}

\item{legend}{A boolean variable that indicates if a legend has to be added}

\item{layout}{A possible starting layout as returned by
umap_graph_clustering (defaults to use the one given in \code{umap_obj})}

\item{update_reference}{A boolean variable. If TRUE, the reference will be
updated based on the label with the highest score}

\item{reference}{The reference to update if update_reference = TRUE}

\item{weight_quantile}{The quantile below which the edges will be put to NA}

\item{weights_normalization_coef}{A normalization coefficient to reduce the
width of the edges in the plots. It defaults to the number of nodes of the
network (with a maximum of 1000)}

\item{show_edges}{A boolean variable to define if the plot should show edges
of the network behind the plot}

\item{only_best_association}{A boolean variable to decide if only the best
association is returned}
}
\value{
A list with the distance scores of the new, labelled clusters and the
old clusters, plus the best association and its confidence and the UMAP plot.
If update_reference=TRUE, also the updated reference
}
\description{
Given a UMAP object umap_obj, obtained from the function
add_to_annotated_reference (the New object), and names for new_clusters,
it computes the diffusion of the new_clusters labels in the UMAP network
defined in umap_obj. If a col_vector is given, the plot will be coloured
as such. If new_points_col is given, the points corresponding to
new_clusters will be plotted twice as big and coloured as such.
If legend=TRUE, a legend will be displayed instead of the labels.
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
res <- add_to_annotated_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
umap_obj <- res
new_clusters <- rownames(new_M)
labelPrediction(umap_obj, new_clusters)
}
