% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_assignment_matrix.R
\name{list_to_assignment_matrix}
\alias{list_to_assignment_matrix}
\title{It returns a membership matrix for each element in the input list}
\usage{
list_to_assignment_matrix(InList)
}
\arguments{
\item{InList}{An input list}
}
\value{
A sparse membership matrix
}
\description{
It returns a membership matrix for each element in the input list
}
\examples{
list_to_assignment_matrix(list('A' = c(1,2,3),
'B' = c(4,3,5),
'C' = c(2,3,5)))
}
