% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_group_contrasts_for_one.R
\name{make_group_contrasts_for_one}
\alias{make_group_contrasts_for_one}
\title{Make group contrasts}
\usage{
make_group_contrasts_for_one(groupX, groups)
}
\arguments{
\item{groupX}{The wanted group}

\item{groups}{All the groups}
}
\value{
A contrast
}
\description{
Make group contrasts
}
\examples{
groupX <- 'A'
groups <- c('A', 'B', 'C')
make_group_contrasts_for_one(groupX, groups)
}
