% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters_to_representation.R
\name{map_clusters_to_representation}
\alias{map_clusters_to_representation}
\title{Wrap-up function to compute a series of summary information given an expression
matrix and a membership vector}
\usage{
map_clusters_to_representation(
  M,
  group,
  BOGs = FALSE,
  pBOGs = FALSE,
  rEEA = FALSE,
  pBOGsig = FALSE
)
}
\arguments{
\item{M}{An expression matrix}

\item{group}{A membership vector}

\item{BOGs}{A boolean variable to define if the BOGs will be computed}

\item{pBOGs}{A boolean variable to define if the pBOGs will be computed}

\item{rEEA}{A boolean variable to define if the rEEA will be computed}

\item{pBOGsig}{A boolean variable to define if the pBOGsig will be computed}
}
\value{
A list with the profiles (get_column_group_average),
signatures (FC_signatures) and bag-of-genes
(run_wilcox_differential + filter_dge_out), the differential expression
table (run_wilcox_differential), the pBOG (FC_signatures_exclusiveSets) and
the pBOGsig (FC_signatures_exclusiveSets)
}
\description{
Wrap-up function to compute a series of summary information given an expression
matrix and a membership vector
}
\examples{
map_clusters_to_representation(M = matrix(seq(1,100), ncol=10),
group = c(rep(c('A','B','C'), each = 3), 'D'))
}
