% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapData.R
\name{mapData}
\alias{mapData}
\title{Map new data onto a reference network}
\usage{
mapData(
  reference_df,
  new_profiles = NULL,
  new_signatures = NULL,
  reference_signatures,
  color_attr,
  label_attr = NULL,
  reference_layout = NULL,
  col_vector = NULL,
  ...
)
}
\arguments{
\item{reference_df}{The annotated reference correlation data frame}

\item{new_profiles}{The profiles that have to be mapped (can directly provide
signatures)}

\item{new_signatures}{The signatures that have to be mapped (can be computed
from profiles)}

\item{reference_signatures}{The reference signatures used to build the
reference network}

\item{color_attr}{The annotation label wanted to be displayed and used
for the annotation of the new points}

\item{label_attr}{The labels to be used in the UMAP. If NULL, the labels of
color_attr will be used, otherwise color_attr will be used to
color the clusters and label_attr to show the labels.}

\item{reference_layout}{The reference network layout object}

\item{col_vector}{The color vector palette. Should be a named vector where
names correspond to the unique values in the color_attr vector.}

\item{...}{Additional parameters given to \code{add_to_reference}}
}
\value{
The umap object and plot before and after adding the new clusters,
plus mapping quality metrics and nearest neighbor annotation
}
\description{
Map new data onto a reference network
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
mapData(reference_df = annotated_M,
new_signatures = new_clusterS,
reference_signatures = refS,
color_attr = 'Best.Assignment')
}
