% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenes2.R
\name{mapGenes2}
\alias{mapGenes2}
\title{Creates a gene-view over a cluster-view UMAP, in which the most impactful genes
for the layout of the cluster-view UMAP are displayed}
\usage{
mapGenes2(
  umap_obj,
  refSignatures,
  color_attr,
  new_clusters = NULL,
  col_vector = NULL,
  n_genes = 100,
  new_points_col = "#FF0000",
  gene_set = NULL,
  threshold = 0.95,
  proportional = TRUE,
  title = NULL
)
}
\arguments{
\item{umap_obj}{A UMAP object as given by the add_to_annotated_reference
function (but specifying 'New').
It has to be an object that contains the following chain:
\code{umap_obj$umap_out$layout}}

\item{refSignatures}{The signatures of the clusters of the UMAP (excluding
new clusters)}

\item{color_attr}{The annotation labels to be used in the UMAP}

\item{new_clusters}{The names of the new clusters}

\item{col_vector}{The color palette to be used in the UMAP}

\item{n_genes}{The number of genes that have to be displayed}

\item{new_points_col}{The color of the new clusters in the UMAP}

\item{gene_set}{A vector from which to choose the displayed genes (if NULL,
all genes are considered)}

\item{threshold}{A threshold to keep only the most significant genes, even
if it might lead to less than n_genes to be displayed.}

\item{proportional}{A boolean variable that defines if the genes to be
displayed are proportional in number to the number of clusters in each color_attr
defined in color_attr. If FALSE, each color_attr in color_attr will
have approximately the same number of genes displayed}

\item{title}{The title of the plot}
}
\value{
A SimpleList with the original cluster-view and the new gene-view,
along with a list of the displayed genes for each color_attr
}
\description{
Creates a gene-view over a cluster-view UMAP, in which the most impactful genes
for the layout of the cluster-view UMAP are displayed
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(stats::runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(stats::runif(100,0.1,7), ncol = 5))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(stats::runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
to_reference <- add_to_annotated_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
mapGenes2(umap_obj = to_reference$New,
new_clusters = rownames(new_M),
refSignatures = S,
color_attr = annotated_M$`Best.Assignment`,
threshold = 0.05)
}
