% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNetwork.R
\name{mapNetwork}
\alias{mapNetwork}
\title{Map network}
\usage{
mapNetwork(
  net,
  new_profiles,
  color_attr = "SubClass",
  label_attr = "SubClass",
  col_vector = "SubClass_colors",
  new_name = NULL,
  n_nearest = 15,
  ...
)
}
\arguments{
\item{net}{The reference network}

\item{new_profiles}{The new profiles to map}

\item{color_attr}{The color attribute in \code{net}, defaults to \code{SubClass}}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}

\item{col_vector}{The color vector in \code{net}, defaults to \code{SubClass_colors}}

\item{new_name}{A vector to define the names of the new profiles. Defaults to
their colnames}

\item{n_nearest}{The number of nearest neighbors to consider for the
layout definition, mapping quality assessment and annotation. Defaults to 15}

\item{...}{Other parameters for the \code{plotSameLayout} function}
}
\value{
An \code{S4Vectors} list containing the correlation matrix between the
profiles and the reference pseudobulks (\code{new_cor}), a data.frame containing
the edges information to connect the new profiles with the pseudobulks in the
network (\code{new_edges}), a matrix containing the layout information, including
the x and y coordinates (\code{new_layout}), the new plot (\code{new_plot}), a mapping
quality assessment (if mapping_quality = TRUE) (\code{mapping_quality}), and
the annotation scores (if annotate = TRUE) (\code{annotation})
}
\description{
Map network
}
