% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_split.R
\name{matrix_split}
\alias{matrix_split}
\title{Splits a matrix in submatrices based on a membership vector}
\usage{
matrix_split(M, group)
}
\arguments{
\item{M}{A matrix}

\item{group}{A membership vector for the columns of the matrix}
}
\value{
A list of submatrices
}
\description{
Splits a matrix in submatrices based on a membership vector
}
\examples{
matrix_split(matrix(sample(seq(0,1),100, replace=TRUE), ncol=10),
group=c(rep(c(1,2,3), each = 3), 4))
}
