% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_normalize.R
\name{min_max_normalize}
\alias{min_max_normalize}
\title{Min-Max normalization}
\usage{
min_max_normalize(vector)
}
\arguments{
\item{vector}{A numerical vector}
}
\value{
The normalized numerical vector with a new distribution between 0 and
1
}
\description{
Min-Max normalization
}
\examples{
v <- c(1,6,5,3,7,8,3,2)
n_v <- neuRoDev:::min_max_normalize(v)
}
