% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_intersect.R
\name{mult_intersect}
\alias{mult_intersect}
\title{Multiple intersection}
\usage{
mult_intersect(vec_list)
}
\arguments{
\item{vec_list}{A vector list}
}
\value{
Common elements between all vectors in the list
}
\description{
Multiple intersection
}
\examples{
v1 <- c('Z', 'A', 'B', 'D')
v2 <- c('A', 'B', 'C', 'D')
v3 <- c('B', 'Z', 'D', 'H')
mult_intersect(list(v1, v2, v3))
}
