% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_annotation.R
\name{nearest_neighbor_annotation}
\alias{nearest_neighbor_annotation}
\title{Nearest neighbor annotation}
\usage{
nearest_neighbor_annotation(
  reference_df,
  new_clusters,
  umap_obj,
  color_attr,
  col_vector = NULL,
  n_nearest = NULL,
  title = NULL,
  to_exclude = NULL,
  compute_means = FALSE
)
}
\arguments{
\item{reference_df}{The reference correlation matrix.}

\item{new_clusters}{The new clusters names.}

\item{umap_obj}{An already computed umap object as returned by
umap_graph_clustering}

\item{color_attr}{The labels to use for the annotation}

\item{col_vector}{A named color palette vector for the annotations}

\item{n_nearest}{The number of nearest neighbors to consider. It defaults
to \code{((dim(distance_matrix)[2])-1)}, where distance_matrix is given by:
\code{umap_obj$umap_out$knn$distances} or
\code{umap_obj$umap_out$refined_network$distances}}

\item{title}{The title to add to the plots}

\item{to_exclude}{The labels inside \code{reference_df[,color_attr]}
that the user doesn't want to consider.}

\item{compute_means}{A boolean variable. If TRUE, it computes the mean instead
of the sum of values belonging to a given group.}
}
\value{
A list that contains, for each new cluster, the annotations of the
neighbors, the most frequent annotation, the highest value in terms of
weighted frequency (based on the distance of the neighbor), a Heatmap and a
barplot and the final number of nearest neighbors considered. If a sub
annotation is provided, the same will be given for the sub annotation as well
}
\description{
Given an annotated reference dataframe reference_df,
a set of new clusters to analyze in new_clusters, a umap object as run
with add_to_reference, which contains the new clusters to
analyze, and a given annotation label (and possibly a
sub_color_attr), it returns the annotation (and sub_annotation) of
each cluster based on its nearest neighbors. With n_nearest you can set
how many nearest neighbours to consider (it will consider first the
closest ones). You can provide a col_vector and sub_col_vector palettes
for the final barplots (otherwise it will create a palette).
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
res <- add_to_reference(annotated_M,
new_M,
annotated_M$`Best.Assignment`)
umap_obj <- res$New
new_clusters <- rownames(new_M)
nearest_neighbor_annotation(annotated_M,
new_clusters,
umap_obj,
color_attr = 'Best.Assignment')
}
