% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_smoothing.R
\name{network_smoothing}
\alias{network_smoothing}
\title{Network smoothing}
\usage{
network_smoothing(network, scores, n_nearest_smooth = NULL)
}
\arguments{
\item{network}{The igraph network}

\item{scores}{The nodes scores to smooth}

\item{n_nearest_smooth}{The number of nearest neighbors to consider for
smoothing. Defaults to the number of nodes.}
}
\value{
The smoothed scores
}
\description{
Network smoothing
}
\examples{
set.seed(123)
g <- igraph::erdos.renyi.game(n = 10,
p.or.m = 0.3,
type = "gnp",
directed = FALSE)
igraph::E(g)$weight <- sample(seq(1,10),
length(igraph::E(g)),
replace = TRUE)
scores <- stats::runif(igraph::vcount(g), min = 0, max = 100)
neuRoDev:::network_smoothing(g, scores)
}
