% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eMatrix.R
\name{plot_eMatrix}
\alias{plot_eMatrix}
\title{Title}
\usage{
plot_eMatrix(net, genes, nRand = 100, eMatrix = NULL)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the expression enrichment}

\item{nRand}{The number of random sampling to use (defaults to 100)}

\item{eMatrix}{If an eMatrix was already computed, it can be given here. If
NULL (default), it is computed.}
}
\value{
An Heatmap of the eMatrix
}
\description{
Title
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_color <- stages_palette[net$Stages]
plot_eMatrix(net = net, genes = rownames(net)[seq(1,5)])
}
