% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eTrace.R
\name{plot_eTrace}
\alias{plot_eTrace}
\title{Plot the eTraces}
\usage{
plot_eTrace(
  net,
  genes = NULL,
  score = NULL,
  expression_enrichment = FALSE,
  main = NULL,
  upper_colors = NULL,
  lower_colors = NULL,
  nRand = NULL,
  ylab = "score",
  pval_threshold = 0.05,
  add_sig_line = FALSE,
  clusters_comparison = FALSE,
  show_ages = FALSE,
  n_ages = 8
)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the expression enrichment}

\item{score}{A score to plot on the y axis}

\item{expression_enrichment}{A boolean, if TRUE the expression enrichment is
computed and plotted (\code{get_eTrace} function). Defaults to FALSE}

\item{main}{The title of the plot}

\item{upper_colors}{The colors for the upper plot}

\item{lower_colors}{The colors for the lower plot}

\item{nRand}{The number of random sampling to use. If NULL, a nRand
proportional to the number of genes is computed (the higher the number of
genes, the lower the number of nRand computed, to a minimum of 200).}

\item{ylab}{The ylab to add. Defaults to \code{score}}

\item{pval_threshold}{A p-value threshold to consider a value to be
significant or not.}

\item{add_sig_line}{A boolean variable. If TRUE, dashed lines are added at
enrichment scores corresponding to a p-value of \code{pval_threshold}, one for
positive values, one for negative values. Values above (or below) those lines
are statistically significant. Defaults to FALSE. If TRUE,
\code{expression_enrichment} is set to TRUE by default.}

\item{clusters_comparison}{A boolean variable. If TRUE, it highlights
clusters that belong to combinations of subclasses and stages that are
significantly higher that the other combinations of subclasses and stages
both overall and within each stage. Defaults to FALSE. If TRUE,
\code{expression_enrichment} is set to TRUE by default.}

\item{show_ages}{A boolean variable. If TRUE, it shows on the x-axis ages
labels.}

\item{n_ages}{If show_ages is TRUE, it defines how many labels to show.
Defaults to 8.}
}
\value{
The plot of the eTraces divided into two subplots
}
\description{
Plot the eTraces
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_color <- subclass_palette[net$SubClass]
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_color <- stages_palette[net$Stages]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from,
'to' = edges_to,
'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
SummarizedExperiment::rowData(net)$informative <- sample(c(TRUE, FALSE), size = nrow(net),
replace = TRUE)
plot_eTrace(net, 'Gene-1')
}
