% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geneset_network.R
\name{plot_geneset_network}
\alias{plot_geneset_network}
\title{Plot the annotation network}
\usage{
plot_geneset_network(
  net,
  quantile_to_remove = NULL,
  increase = 1,
  remove_unconnected = TRUE
)
}
\arguments{
\item{net}{The annotation network}

\item{quantile_to_remove}{Quantile of the edges weight distribution under
which edges are removed}

\item{increase}{The increase coefficient for the size of the edges}

\item{remove_unconnected}{A boolean, if TRUE unconnected nodes (after
edges removal) are removed.}
}
\value{
The igraph plot of the network
}
\description{
Plot the annotation network
}
\examples{
net <- igraph::sample_k_regular(10, 3, directed = FALSE, multiple = FALSE)
igraph::E(net)$weight <- sample(seq(1,10), length(igraph::E(net)),
replace = TRUE)
igraph::V(net)$size <- sample(seq(1,10), length(igraph::V(net)),
replace = TRUE)
plot_geneset_network(net)
}
