% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{Create a ggplot of the network}
\usage{
plot_net(
  layout,
  color_attr,
  edges = NULL,
  label_attr = NULL,
  title = NULL,
  new_points_col = "#FF0000",
  legend = FALSE,
  legend_name = NULL,
  only_new_points = FALSE,
  alpha = NULL,
  show_edges = TRUE,
  n_increase = 1,
  stroke = 0.2
)
}
\arguments{
\item{layout}{A data frame with the layout informations to plot}

\item{color_attr}{The color attribute}

\item{edges}{The edges to visualize}

\item{label_attr}{A label vector for the clusters}

\item{title}{A title}

\item{new_points_col}{The color of newly mapped points}

\item{legend}{A boolean, if TRUE the legend is shown. Defaults to FALSE}

\item{legend_name}{The legend name}

\item{only_new_points}{A boolean, if TRUE only new points are labeled. Defaults
to FALSE}

\item{alpha}{A vector with the alpha information. Defaults to all 1}

\item{show_edges}{A boolean, if FALSE no edges are shown}

\item{n_increase}{The increase for newly mapped points. Defaults to 1}

\item{stroke}{The stroke around the points, defaults to 0.2}
}
\value{
A ggplot
}
\description{
Create a ggplot of the network
}
