% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesRank.R
\name{plotGenesRank}
\alias{plotGenesRank}
\title{Plot the rank of a gene (set) across the UMAP}
\usage{
plotGenesRank(
  umap_obj,
  genes,
  refExpression,
  newExpression = NULL,
  label_attr = NULL,
  col_vector = NULL,
  new_clusters = NULL,
  title = NULL,
  smooth = TRUE,
  n_nearest_smooth = NULL,
  relative = TRUE,
  n_components = 2,
  weight_quantile = 0.75,
  weights_normalization_coef = NULL,
  show_edges = TRUE,
  rank = FALSE,
  na.vec = NULL
)
}
\arguments{
\item{umap_obj}{A UMAP object as given by the umap_signature_plot function or
by the add_to_annotated_reference function.
It has to be an object that contains the following chain:
\code{umap_obj$umap_obj$umap_out$layout}}

\item{genes}{A vector of one or multiple genes whose ranks have to be plotted}

\item{refExpression}{The expression matrix of the clusters in the UMAP}

\item{newExpression}{The expression matrix of the new clusters}

\item{label_attr}{The labels that are wanted in the new UMAP}

\item{col_vector}{A color vector for the UMAP. If NULL, different shades of
blues9 will be given (the darker the color, the higher the rank of the genes,
so the higher the signature)}

\item{new_clusters}{If the umap_obj contains also new clusters (if originated
from add_to_annotated_reference, 'New'), and this new clusters want to be
plotted bigger in the UMAP, their names have to be specified here}

\item{title}{A title to give. If NULL, the names of the genes will be plotted}

\item{smooth}{A boolean variable to define if a smoothing step will be
performed to define the scores of each cluster}

\item{n_nearest_smooth}{The number of nearest neighbors to use to smooth the
rank value of expression. Defaults to all nodes.}

\item{relative}{A boolean variable to define if the colors should be relative
or absolute. If smooth, the result will be relative regardless of the input}

\item{n_components}{The number of dimensions to use in the plot}

\item{weight_quantile}{The quantile below which the edges will be put to NA}

\item{weights_normalization_coef}{A normalization coefficient to reduce the
width of the edges in the plots. It defaults to the number of nodes of the
network (with a maximum of 1000)}

\item{show_edges}{A boolean variable to define if edges have to be shown or not}

\item{rank}{A boolean variable to define if the values are to be ranked}

\item{na.vec}{A vector with length nrow(refExpression) and with NA where you
want the plot to be transparent and without NAs where you want to see the
expression}
}
\value{
The UMAP with colors and transparency proportional to the ranks of
the genes
}
\description{
Plot the rank of a gene (set) across the UMAP
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
to_reference <- add_to_annotated_reference(annotated_M, new_M, rownames(annotated_M))
plotGenesRank(umap_obj = to_reference$New,
genes = 'Gene-8',
refExpression = S,
newExpression = new_clusterS,
new_clusters = rownames(new_M),
label_attr = colnames(S))
}
