% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot the network}
\usage{
plotNetwork(
  net,
  color_attr = "SubClass",
  label_attr = "SubClass",
  col_vector = NULL,
  no_label = FALSE,
  show_edges = TRUE,
  max_size = 3,
  ...
)
}
\arguments{
\item{net}{The reference network}

\item{color_attr}{The color attribute in \code{net}, defaults to \code{SubClass}}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}

\item{col_vector}{The color vector in \code{net}, defaults to \code{SubClass_colors}}

\item{no_label}{A boolean, if TRUE the clusters' labels will not be shown.
Defaults to FALSE}

\item{show_edges}{A boolean, if FALSE it doesn't show the edges. Defaults to
TRUE}

\item{max_size}{The maximum size of the dots in the network visualization}

\item{...}{Additional parameters for \code{plot_net}}
}
\value{
A ggplot of the network
}
\description{
Plot the network
}
