% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkScore.R
\name{plotNetworkScore}
\alias{plotNetworkScore}
\title{Plot a score onto the network}
\usage{
plotNetworkScore(
  net,
  genes = NULL,
  score = NULL,
  label_attr = NULL,
  smooth = FALSE,
  n_nearest = 15,
  normalize = TRUE,
  na.vec = NULL,
  palette = grDevices::blues9,
  title = NULL,
  show_edges = TRUE,
  stroke = 0.5,
  fix_alpha = FALSE,
  max_size = 3
)
}
\arguments{
\item{net}{The reference network}

\item{genes}{The genes to consider for the score. If multiple genes are given,
the score will be their average.}

\item{score}{The score to plot. If NULL (default), the logcounts of the
pseudobulks are used}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}

\item{smooth}{A boolean, if TRUE (default), the scores are smoothed to consider
the network structure}

\item{n_nearest}{The number of nearest neighbors to consider for smoothing.
Defaults to 15}

\item{normalize}{A boolean, if TRUE (default) and score is NULL, it normalizes
the pseudobulk expression computing a row-wise z-score}

\item{na.vec}{The vector to define if certain clusters are not to be shown}

\item{palette}{The color palette to use. Defaults to \code{blues9}}

\item{title}{The title of the plot}

\item{show_edges}{A boolean, if FALSE no edges are shown. Defaults to TRUE}

\item{stroke}{The stroke of the points, defaults to 0.5}

\item{fix_alpha}{A boolean, if TRUE all points will have alpha 1, otherwise
it will be proportional to the score. Defaults to FALSE}

\item{max_size}{The maximum size of the dots in the network visualization}
}
\description{
Plot a score onto the network
}
