% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSameLayout.R
\name{plotSameLayout}
\alias{plotSameLayout}
\title{Plot a mapped dataset conserving the same layout as the original reference
network}
\usage{
plotSameLayout(
  net,
  new_cor,
  color_attr = "SubClass",
  label_attr = "SubClass",
  col_vector = "SubClass_colors",
  no_label = FALSE,
  new_points_col = "#FF0000",
  max_size = 3,
  only_new_points = FALSE,
  mapping_quality = TRUE,
  annotate = TRUE,
  n_nearest = 15,
  new_name = NULL,
  ...
)
}
\arguments{
\item{net}{The reference network}

\item{new_cor}{The correlation matrix between the profiles and the reference
pseudobulks}

\item{color_attr}{The color attribute in \code{net}, defaults to \code{SubClass}}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}

\item{col_vector}{The color vector in \code{net}, defaults to \code{SubClass_colors}}

\item{no_label}{A boolean, if TRUE the clusters' labels will not be shown.
Defaults to FALSE}

\item{new_points_col}{The color of the mapped points}

\item{max_size}{The maximum size of the dots in the network visualization}

\item{only_new_points}{A boolean, if TRUE only the labels of the new points
are shown. Defaults to FALSE}

\item{mapping_quality}{A boolean, if TRUE (default), it returns also the
mapping quality assessment}

\item{annotate}{A boolean, if TRUE (default), it returns also the
annotation scores for the mapped points}

\item{n_nearest}{The number of nearest neighbors to consider for the
layout definition, mapping quality assessment and annotation. Defaults to 15}

\item{new_name}{A vector to define the names of the new profiles. Defaults to
their colnames}

\item{...}{Additional parameters for \code{plot_net}}
}
\value{
An \code{S4Vectors} list containing the correlation matrix between the
profiles and the reference pseudobulks (\code{new_cor}), a data.frame containing
the edges information to connect the new profiles with the pseudobulks in the
network (\code{new_edges}), a matrix containing the layout information, including
the x and y coordinates (\code{new_layout}), the new plot (\code{new_plot}), a mapping
quality assessment (if mapping_quality = TRUE) (\code{mapping_quality}), and
the annotation scores (if annotate = TRUE) (\code{annotation})
}
\description{
Plot a mapped dataset conserving the same layout as the original reference
network
}
