% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrends.R
\name{plotTrends}
\alias{plotTrends}
\title{plotTrends}
\usage{
plotTrends(
  net,
  pref_exp_genes,
  sce,
  profiles,
  coldata,
  subclass = NULL,
  ylim = NULL,
  together = FALSE,
  relative = FALSE
)
}
\arguments{
\item{net}{The reference network}

\item{pref_exp_genes}{The preferentially expressed genes list}

\item{sce}{The external object in the form of a SingleCellExperiment object}

\item{profiles}{The actual expression matrix used for the mapping}

\item{coldata}{Where to find the information of the replicates used for
averaging in the \code{colData} of \code{sce}}

\item{subclass}{If indicated as a vector it selects only a subset of
reference subclasses. Defaults to all subclasses}

\item{ylim}{The lower and upper limits of the y-axis
(defaults to NULL, which goes back to the automatic R plot ylim)}

\item{together}{Defines if the different preferentially expressed genes
should be plotted together (adjust \code{ylim} accordingly) (defaults to FALSE)}

\item{relative}{If TRUE, the expression is scaled from 0 to 1
(defaults to FALSE)}
}
\value{
A plot of the expression of preferentially expressed genes in the
samples of \code{sce}
}
\description{
plotTrends
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_color <- subclass_palette[net$SubClass]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from, 'to' = edges_to, 'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
pref_exp_genes <- list('A' = c('Gene-1', 'Gene-2'),
'B' = c('Gene-3', 'Gene-4'),
'C' = c('Gene-5', 'Gene-6'),
'D' = c('Gene-7', 'Gene-8'))
random_m <- matrix(sample(seq(1,10, length.out=10000), 15000*20, replace = TRUE), ncol = 20)
rownames(random_m) <- paste0('Gene-', seq(1,15000))
colnames(random_m) <- paste0('Sample-', seq(1,20))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = random_m))
sce$replicates <- rep(c('Rep1', 'Rep2', 'Rep3', 'Rep4'), each = 5)
profiles <- neuRoDev:::get_column_group_average(SingleCellExperiment::logcounts(sce),
sce$replicates)
coldata <- 'replicates'
neuRoDev:::plotTrends(net, pref_exp_genes, sce, profiles, coldata)
}
