% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportional_sampling.R
\name{Proportional_sampling}
\alias{Proportional_sampling}
\title{Proportional sampling}
\usage{
Proportional_sampling(M, membership_vector, ReturnSCE = TRUE, Ntotal = 2000)
}
\arguments{
\item{M}{An expression matrix}

\item{membership_vector}{A membership vector for the columns of the matrix}

\item{ReturnSCE}{A boolean variable that defines if the subsampled matrix
is returned or if only the indexes}

\item{Ntotal}{The number of cells in total}
}
\value{
Either the matrix filtered or a list of indexes and labels
}
\description{
Proportional sampling
}
\examples{
set.seed(123)
Proportional_sampling(matrix(runif(20000,0,10), ncol=200),
membership_vector = rep(c('A','B','C','D'),
each = 50),
Ntotal=20)
}
