% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_based_eea.R
\name{rank_based_eea}
\alias{rank_based_eea}
\title{Ranked based differential expression analysis}
\usage{
rank_based_eea(M, group, FCtresh = 1)
}
\arguments{
\item{M}{An expression matrix}

\item{group}{A membership vector for the columns of M}

\item{FCtresh}{A threshold of the Fold-Change}
}
\value{
A list containing the fraction of non-zero values for each group
(get_column_group_detection_rate), the ranks of rows based on such fraction,
the relative ranks compared to other groups (RelRanks), and the same but when
done pairwise (RelRanks_pairs), the Fold-Change row sum for each gene in
each group (RelRanks_pairs_sum) and the genes with an higher Fold-Change than
FCtresh compared to all the other groups (Gs).
}
\description{
It ranks rows in matrix M based on the number of non-zero
entries they have in members of a group in group (done for each group).
Then computes the differences in (transformed) ranks between each group
and the mean of the others (RelRanks) and the same done pairwise
(RelRanks_pairs). Then, given a Fold-Change threshold of FCtresh, it
returns the genes that have an higher Fold-Change than FCtresh compared
to all the other groups (done for each group) (Gs). Eventually, it also
computes the Fold-Change rowSum for each gene in each group
(RelRanks_pairs_sum).
}
\examples{
rank_based_eea(matrix(sample(seq(0,1),100,
replace=TRUE),
ncol=10),
group=c(rep(c(1,2,3),
each = 3),
4))
}
