% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSignatures.R
\name{rankSignatures}
\alias{rankSignatures}
\title{Ranks a vector (useful for signatures)}
\usage{
rankSignatures(signature)
}
\arguments{
\item{signature}{A signature vector}
}
\value{
The ranks of the genes in the signature
}
\description{
Ranks a vector (useful for signatures)
}
\examples{
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
rankedS <- apply(S, 2, rankSignatures)

}
