% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_signatures_correlation.R
\name{reference_signatures_correlation}
\alias{reference_signatures_correlation}
\title{Reference signatures correlation}
\usage{
reference_signatures_correlation(S, refS, warn = TRUE)
}
\arguments{
\item{S}{The new signatures (a list or a matrix)}

\item{refS}{Reference signatures (a list or a matrix) with the same rownames as S}

\item{warn}{A boolean (defaults to TRUE) that specifies whether a warning
should be given in case the signatures are too similar within them and thus
the results should be interpreted carefully.}
}
\value{
A dataframe that contains the correlation values, the highest correlation
value for each cluster/row of S, the best label (based on highest correlation
value). In case the refS is a list, the highest value and best label are also
specified for each element of the list.
}
\description{
Reference signatures correlation
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
reference_signatures_correlation(S, refS)
}
