% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zero_pvals.R
\name{remove_zero_pvals}
\alias{remove_zero_pvals}
\title{Remove pvalues equal to 0}
\usage{
remove_zero_pvals(xVec)
}
\arguments{
\item{xVec}{A vector containing pvalues}
}
\value{
The vector with the modified 0 pvalues
}
\description{
This function removes the pvalues equal to zeros and replaces them with a pvalue
equal to the minimum pvalue (excluding those equal to 0) times 0.1
}
\examples{
remove_zero_pvals(c(rep(0, 5), seq(0.01,1,0.1)))
}
