% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_group_for_score.R
\name{run_one_group_for_score}
\alias{run_one_group_for_score}
\title{Perform the enrichment for one group at a time}
\usage{
run_one_group_for_score(
  expression_matrix,
  groups,
  groupX,
  Datatype = "RNAseq",
  threshold = 0.1
)
}
\arguments{
\item{expression_matrix}{The expression matrix to use}

\item{groups}{The groups in which to devide the expression matrix}

\item{groupX}{The wanted group}

\item{Datatype}{The expression_matrix type. Defaults to RNAseq.}

\item{threshold}{The threshold for the pvalue above which to put the coef
returned by limma::eBayes to 0.}
}
\value{
The sum of the coefficients for the comparison between groupX and the other
groups.
}
\description{
Perform the enrichment for one group at a time
}
\examples{
expression_matrix <- matrix(sample(seq(1,100), 10000, replace = TRUE), ncol = 10)
rownames(expression_matrix) <- paste0('Gene.', seq(1, nrow(expression_matrix)))
colnames(expression_matrix) <- paste0('Column.', seq(1, ncol(expression_matrix)))
groups <- paste0('Group.', c(1,1,2,3,3,4,5,4,2,5))
groupX <- 'Group.1'
run_one_group_for_score(expression_matrix, groups, groupX)
}
