% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seurat_clustering.R
\name{run_seurat_clustering}
\alias{run_seurat_clustering}
\title{Run standard Seurat pipeline given a matrix}
\usage{
run_seurat_clustering(
  expression_matrix,
  PCs_to_use = NULL,
  verbose = FALSE,
  resolution = 1
)
}
\arguments{
\item{expression_matrix}{An expression matrix}

\item{PCs_to_use}{The number of Principal Components to consider for
FindNeighbors of Seurat}

\item{verbose}{A boolean variable to define if the steps of Seurat will be
printed or not}

\item{resolution}{The resolution to consider for FindClusters of Seurat}
}
\value{
A seurat object with clusters, UMAP, PCA and the normalized counts
}
\description{
Run standard Seurat pipeline given a matrix
}
\examples{
set.seed(123)
M <- matrix(runif(200000,0,100), ncol=2000)
rownames(M) <- paste0('Gene-', seq(1, dim(M)[1]))
colnames(M) <- paste0('Cell-', seq(1, dim(M)[2]))
run_seurat_clustering(M)
}
