% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wilcox_differential.R
\name{run_wilcox_differential}
\alias{run_wilcox_differential}
\title{Wilcoxon differential expression analysis}
\usage{
run_wilcox_differential(expMat, binGroup, orderOut = FALSE)
}
\arguments{
\item{expMat}{A genetic expression matrix}

\item{binGroup}{A numerical membership vector for the columns of expMat}

\item{orderOut}{A boolean variable to order the output based on the AUC
(decreasing = TRUE)}
}
\value{
It returns a dataframe of the differential expression analysis with
the main statistics, like the p-value of each gene, the Bonferroni adjusted
pvalue (padj), the log Fold-Change, the AUC, the average expression (avgExpr),
and a score which is the -log10 of the adjusted
pvalue multiplied by the sign of the logFC.
}
\description{
This function computes the Wilcoxon test of genetic expression matrices given
a membership vector.
}
\examples{
run_wilcox_differential(expMat = matrix(seq(1,200), ncol=10),
binGroup = c(rep(c(1,2,3), each = 3), 4))
}
