% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMapping.R
\name{scoreMapping}
\alias{scoreMapping}
\title{Mapping scores}
\usage{
scoreMapping(net, new_cor, label_attr = "SubClass")
}
\arguments{
\item{net}{The reference network}

\item{new_cor}{The correlation matrix between the profiles and the reference
pseudobulks}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}
}
\value{
An S4Vectors list containing the cluster accuracy \code{ClusterAccuracy},
the label precision (\code{LabelPrecision}), the stage precision (\code{StagePrecision}),
the global accuracy (\code{GlobalAccuracy}), the global precision (\code{GlobalPrecision}),
the mapping score (\code{MappingScore})
}
\description{
Mapping scores
}
