% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sce_cols.R
\name{split_sce_cols}
\alias{split_sce_cols}
\title{Split a SingleCellExperiment object by column}
\usage{
split_sce_cols(sce, colVec)
}
\arguments{
\item{sce}{The SingleCellExperiment}

\item{colVec}{A vector that groups columns}
}
\value{
A list of SingleCellExperiment objects with one element per group
}
\description{
Split a SingleCellExperiment object by column
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
neuRoDev:::split_sce_cols(net, net$SubClass)
}
