% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_graph_clustering.R
\name{umap_graph_clustering}
\alias{umap_graph_clustering}
\title{Computes the UMAP given a matrix}
\usage{
umap_graph_clustering(
  M,
  method = "louvain",
  resolution = NULL,
  n_neighbors = 15,
  transpose = TRUE,
  n_components = 2,
  refine_network = TRUE,
  weight_quantile = NULL,
  weights_normalization_coef = 1
)
}
\arguments{
\item{M}{A matrix}

\item{method}{A method for clustering an igraph network}

\item{resolution}{A resolution parameter in case the method is leiden.
Defaults to 1}

\item{n_neighbors}{The number of neighbors to consider in the UMAP}

\item{transpose}{A boolean variable that defines if the M matrix has to
be trasposed. Default is TRUE, as the matrix returned by
reference_signatures_correlation has to be transposed.}

\item{n_components}{The number of components, or dimensions, to compute}

\item{refine_network}{A boolean variable (defaults to TRUE) that defines if
a correlation-distance based network has to be computed and visualized in the
2D plot}

\item{weight_quantile}{The quantile below which the edges will be put to NA}

\item{weights_normalization_coef}{A normalization coefficient to reduce the
width of the edges in the plots. It defaults to the number of nodes of the
network (with a maximum of 1000)}
}
\value{
A list containing the UMAP object, the KNN network and the output
of the clustering
}
\description{
Given a matrix M, it computes the UMAP and returns the
object, with the KNN graph, clustered with either Louvain, Leiden or
Walktrap (defined with method)
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
umap_graph_clustering(M)
}
