% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot_same_layout.R
\name{umap_plot_same_layout}
\alias{umap_plot_same_layout}
\title{Plots the UMAP given new clusters, keeping the same layout before and after
adding them}
\usage{
umap_plot_same_layout(
  reference_df,
  signatures_cor,
  color_attr,
  mapping_quality = TRUE,
  annotate = TRUE,
  refine_network = TRUE,
  umap_obj = NULL,
  method = "louvain",
  new_clusters = NULL,
  col_vector = NULL,
  new_name = NULL,
  label_attr = NULL,
  title = NULL,
  new_points_col = "#FF0000",
  legend = FALSE,
  legend_name = NULL,
  only_new_points = FALSE,
  resolution = NULL,
  n_neighbors = 15,
  n_components = 2,
  weight_quantile = NULL,
  weights_normalization_coef = 1,
  show_edges = TRUE,
  n_increase = 0.2,
  alpha = NULL,
  no_label = FALSE,
  ...
)
}
\arguments{
\item{reference_df}{The reference correlation matrix.}

\item{signatures_cor}{The signature correlation matrix as defined in
reference_signatures_correlation}

\item{color_attr}{The groups to show in the UMAP. It can be a column in
signatures_cor or a vector. It should match the names of col_vector, if given.}

\item{mapping_quality}{A boolean to define if the mapping quality should be
computed}

\item{annotate}{A boolean to define if the nearest
neighbor annotation has to be computed}

\item{refine_network}{A boolean variable (defaults to TRUE) that defines if
a correlation-distance based network has to be computed and visualized in the
2D plot}

\item{umap_obj}{An already computed umap object as returned by
umap_graph_clustering}

\item{method}{The method to use in umap_graph_clustering. Can be either
'leiden', 'louvain' (default) or 'walktrap'}

\item{new_clusters}{The new clusters names.}

\item{col_vector}{A named color vector palette with a color for each color_attr}

\item{new_name}{The name/label that has to be given to new clusters.}

\item{label_attr}{The labels to be used in the UMAP. If NULL, the groups given
in color_attr are shown. If no labels are wanted, set no_label to TRUE.}

\item{title}{The title of the UMAP}

\item{new_points_col}{The color of the new clusters in the UMAP plot}

\item{legend}{A boolean variable that indicates if a legend has to be added}

\item{legend_name}{The name of the legend}

\item{only_new_points}{A boolean variable that defines if only the new clusters
are labelled}

\item{resolution}{Resolution parameter for method='leiden'. Defaults to 1}

\item{n_neighbors}{The number of neighbors to be used in umap_graph_clustering.
It defaults to \code{min(15 + dim(signatures_cor)[1], dim(signatures_cor)[1])}}

\item{n_components}{The number of dimensions to use in the plot}

\item{weight_quantile}{The quantile below which the edges will be put to NA}

\item{weights_normalization_coef}{A normalization coefficient to reduce the
width of the edges in the plots. It defaults to the number of nodes of the
network (with a maximum of 1000)}

\item{show_edges}{A boolean variable to define if the plot should show edges
of the network behind the plot}

\item{n_increase}{The increase in size of new edges}

\item{alpha}{The alpha vector for the colors in the UMAP. Defaults to 1}

\item{no_label}{A boolean variable. If TRUE, no label is shown.}

\item{...}{Additional parameters for \code{nearest_neighbor_annotation}}
}
\value{
The umap object and plot before and after adding the new clusters,
plus mapping quality metrics and nearest neighbor annotation
}
\description{
Plots the UMAP given new clusters, keeping the same layout before and after
adding them
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
annotated_M <- reference_signatures_correlation(S, refS)
new_clusterS <- FC_signatures(matrix(runif(80,0,10), ncol = 4))
rownames(new_clusterS) <- paste0('Gene-', seq(1, dim(new_clusterS)[1]))
colnames(new_clusterS) <- paste0('New-', seq(1, dim(new_clusterS)[2]))
new_M <- reference_signatures_correlation(new_clusterS, refS)
umap_plot_same_layout(reference_df = annotated_M,
signatures_cor = new_M,
color_attr = annotated_M$`Best.Assignment`)
}
