% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_pointsize.R
\name{umap_pointsize}
\alias{umap_pointsize}
\title{Given the layout of a UMAP, it plots it}
\usage{
umap_pointsize(
  layout,
  color_attr,
  edges = NULL,
  label_attr = NULL,
  title = NULL,
  new_points_col = "#FF0000",
  legend = FALSE,
  legend_name = NULL,
  only_new_points = FALSE,
  alpha = NULL,
  show_edges = TRUE,
  n_increase = 1
)
}
\arguments{
\item{layout}{In order, the X and Y coordinates, the size, the colors of
each color_attr, the color_attr of each row (has to match the colors),
and in case the sub groups.}

\item{color_attr}{The color_attr variable (usually the same as column 4 of layout)}

\item{edges}{The edges object (an object that contains the following 7
columns: from, to, weight, from.x, from.y, to.x, to.y)}

\item{label_attr}{A label attribute to add to the plot}

\item{title}{A title to the plot}

\item{new_points_col}{If some points want to be highlighted, which color(s)
they have to be}

\item{legend}{A boolean variable to define if a legend is added}

\item{legend_name}{The legend name (if legend=TRUE)}

\item{only_new_points}{A boolean variable if some points want to be
highlighted to define if the labels are added only for such points}

\item{alpha}{A vector of values to be used as transparency}

\item{show_edges}{A boolean variable to determine if the edges of the network
behind the plot should be shown}

\item{n_increase}{The increase in size of new edges}
}
\value{
The UMAP plot
}
\description{
Given the layout of a UMAP, it plots it
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
umap <- umap_graph_clustering(M)
l <- umap$umap_out$layout
l <- cbind(l, 'Size' = rep(1, dim(l)[1]))
l <- cbind(l, 'Color' = rep(c('blue', 'green'), each=5))
l <- cbind(l, 'Group' = rep(c('A','B'), each=5))
umap_pointsize(l, color_attr = rep(c('A','B'), each=5), legend = TRUE)
}
