% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_signature_plot.R
\name{umap_signature_plot}
\alias{umap_signature_plot}
\title{UMAP plot of the signatures}
\usage{
umap_signature_plot(
  signatures_cor,
  refine_network = TRUE,
  color_attr = NULL,
  label_attr = NULL,
  col_vector = NULL,
  umap_obj = NULL,
  title = NULL,
  method = "louvain",
  legend = FALSE,
  legend_name = NULL,
  resolution = NULL,
  n_neighbors = 15,
  n_components = 2,
  weight_quantile = NULL,
  weights_normalization_coef = 1,
  show_edges = TRUE,
  alpha = NULL,
  no_label = FALSE
)
}
\arguments{
\item{signatures_cor}{The signature correlation matrix as defined in
reference_signatures_correlation}

\item{refine_network}{A boolean variable (defaults to TRUE) that defines if
a correlation-distance based network has to be computed and visualized in the
2D plot}

\item{color_attr}{The groups to show in the UMAP. It can be a column in
signatures_cor or a vector. It should match the names of col_vector, if given.}

\item{label_attr}{The labels to be used in the UMAP. If NULL, the groups given
in color_attr are shown. If no labels are wanted, set no_label to TRUE.}

\item{col_vector}{A named color vector palette with a color for each color_attr}

\item{umap_obj}{An already computed umap object as returned by
umap_graph_clustering}

\item{title}{The title of the UMAP}

\item{method}{The method to use in umap_graph_clustering. Can be either
'leiden', 'louvain' (default) or 'walktrap'}

\item{legend}{A boolean variable that indicates if a legend has to be added}

\item{legend_name}{The name of the legend}

\item{resolution}{Resolution parameter for method='leiden'. Defaults to 1}

\item{n_neighbors}{The number of neighbors to be used in umap_graph_clustering.
It defaults to \code{min(15 + dim(signatures_cor)[1], dim(signatures_cor)[1])}}

\item{n_components}{The number of dimensions to use in the plot}

\item{weight_quantile}{The quantile below which the edges will be put to NA}

\item{weights_normalization_coef}{A normalization coefficient to reduce the
width of the edges in the plots. It defaults to the number of nodes of the
network (with a maximum of 1000)}

\item{show_edges}{A boolean variable to determine if the edges of the network
behind the plot should be shown}

\item{alpha}{The alpha vector for the colors in the UMAP. Defaults to 1}

\item{no_label}{A boolean variable. If TRUE, no label is shown.}
}
\value{
A list with the UMAP object and the UMAP plot
}
\description{
Given a matrix of correlation with signatures_cor,
a color_attr parameter, a label attribute label_attr for the plot, it returns
a UMAP object and a UMAP plot. The UMAP object can be given to the
function with umap_obj. The clustering method of the UMAP is Louvain as
default, but it can be also Walktrap ('walktrap') and Leiden ('leiden').
A title for the UMAP can be given. A col_vector palette can be defined, one
color for each color_attr. A legend can be added (legend=TRUE). A resolution
parameter can be defined, if method='leiden'. The number of neighbors for
umap_graph_clustering is defined by n_neighbors.
}
\examples{
set.seed(123)
S <- FC_signatures(matrix(runif(200,0,10), ncol = 10))
rownames(S) <- paste0('Gene-', seq(1, dim(S)[1]))
refS <- FC_signatures(matrix(runif(200,0.1,7), ncol = 10))
colnames(refS) <- paste0('Reference-', seq(1, dim(refS)[2]))
rownames(refS) <- paste0('Gene-', seq(1, dim(refS)[1]))
M <- reference_signatures_correlation(S, refS)
color_attr = M$`Best.Assignment`
umap_signature_plot(M,
color_attr = color_attr,
label_attr = color_attr)
}
