% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_table.R
\name{value_table}
\alias{value_table}
\title{Sum of each element in a group}
\usage{
value_table(
  v,
  perc = FALSE,
  sort = TRUE,
  reciprocal = FALSE,
  compute_means = FALSE
)
}
\arguments{
\item{v}{A named numerical vector. The names define the groups}

\item{perc}{A boolean variable. If TRUE, the result is given as a percentage
out of the total numbers (sum between groups = 1)}

\item{sort}{A boolean variable. If TRUE, results will be sorted
(decreasing=TRUE)}

\item{reciprocal}{A boolean variable. If TRUE, the vector will be first turned
into 1/v}

\item{compute_means}{A boolean variable. If TRUE, it computes the mean instead
of the sum of values belonging to a given group.}
}
\value{
The sum of each element in a group, where the groups are the names of
the vector
}
\description{
Sum of each element in a group
}
\examples{
v <- seq(1,100)
names(v) <- rep(c('A', 'B', 'C', 'D'), each = 25)
neuRoDev:::value_table(v)
neuRoDev:::value_table(v, perc=TRUE)
neuRoDev:::value_table(v, perc=TRUE, reciprocal=TRUE)
}
