% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_summary.R
\name{weighted_summary}
\alias{weighted_summary}
\title{Weighted summary}
\usage{
weighted_summary(v, weights)
}
\arguments{
\item{v}{A numerical vector}

\item{weights}{A numerical vector of weights}
}
\value{
The weighted summary
}
\description{
Weighted summary
}
\examples{
set.seed(123)
weighted_summary(seq(1,100), weights=runif(100,0,1))
}
