% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMapping.R
\name{annotateMapping}
\alias{annotateMapping}
\title{Add annotations to the mapped profiles}
\usage{
annotateMapping(
  net,
  new_cor,
  color_attr = "SubClass",
  col_vector = NULL,
  n_nearest = 15,
  compute_means = FALSE,
  order_names = FALSE
)
}
\arguments{
\item{net}{The reference network}

\item{new_cor}{The correlation matrix between the profiles and the reference
pseudobulks}

\item{color_attr}{The color attribute in \code{net}, defaults to \code{SubClass}}

\item{col_vector}{The color vector in \code{net}, defaults to \code{SubClass_color}}

\item{n_nearest}{The number of nearest neighbors to consider for the
layout definition, mapping quality assessment and annotation. Defaults to 15}

\item{compute_means}{A boolean, if TRUE the means of the neighbors scores are
considered instead of the sum (default)}

\item{order_names}{A boolean, if TRUE the names in annotateMapping
barplot are ordered. Defaults to FALSE}
}
\value{
An \code{S4Vectors} list containing the global confidence of the mapping
annotations (\code{Global.Confidence}), the single annotations confidence
(\code{Annotations.Confidence}), the best annotation per mapped point
(\code{Best.Annotation}), the highest confidence value per mapped point
(\code{Best.Confidence}), the annotations as a percentage (\code{Annotations.Perc}),
and a summary barplot (\code{Barplot})
}
\description{
Add annotations to the mapped profiles
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_color <- subclass_palette[net$SubClass]
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_color <- stages_palette[net$Stages]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from,
'to' = edges_to,
'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
SummarizedExperiment::rowData(net)$informative <- sample(c(TRUE, FALSE), size = nrow(net),
replace = TRUE)
new_profiles <- matrix(sample(seq(1,10, length.out=10000), nrow(net)*10, replace = TRUE), ncol = 10)
rownames(new_profiles) <- rownames(net)
colnames(new_profiles) <- paste0('NewCol-', seq(1,10))
common_genes <- intersect(rownames(net)[SingleCellExperiment::rowData(net)$informative],
rownames(new_profiles))
new_cor <- stats::cor(t(apply(as.matrix(SingleCellExperiment::logcounts(net)[common_genes,]),
1, function(v) {(v-mean(v))/stats::sd(v)})),new_profiles[common_genes,])
annotation <- annotateMapping(net, new_cor)
}
