% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_average_expression.R
\name{geneset_average_expression}
\alias{geneset_average_expression}
\title{Get average expression for genesets or geneset groups}
\usage{
geneset_average_expression(
  geneset_df,
  pseudobulk,
  genesets = "AnnotationGroup",
  celltypes = "Class",
  genes = "TopGenesByGroup"
)
}
\arguments{
\item{geneset_df}{The DF that specifies the wanted genesets}

\item{pseudobulk}{The wanted pseudobulk expression}

\item{celltypes}{The column in which the celltypes are specified in
\code{geneset_df}}

\item{genes}{The column in which the wanted genes are specified in
\code{geneset_df}}

\item{genesets_annotations}{The column in which the genesets names are
specified in \code{geneset_df}. Defaults to "AnnotationGroup"}
}
\value{
The average expression of each pseudobulk column for each geneset
}
\description{
Get average expression for genesets or geneset groups
}
\examples{
anno_df <- S4Vectors::DataFrame('AnnotationGroup' = paste0('AnnoGroup-',
seq(1, 10)),
'Class' = rep('Class1', 10),
'TopGenesByGroup' = rep('Gene.1-Gene.2', 10))
pseudo <- SingleCellExperiment::SingleCellExperiment(
assays = list(logcounts = matrix(sample(seq(1,100),
4*10,
replace = TRUE),
nrow = 4)))
rownames(pseudo) <- paste0('Gene.', seq(1, nrow(pseudo)))
colnames(pseudo) <- paste0('Column-', seq(1, ncol(pseudo)))
geneset_average_expression(anno_df, pseudo)
}
