% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterSignatures.R
\name{getClusterSignatures}
\alias{getClusterSignatures}
\title{Signatures Initialization}
\usage{
getClusterSignatures(
  expression_matrix,
  split = NULL,
  subsample = FALSE,
  PCs_to_use = NULL,
  resolution = 1,
  clusters_mv = NULL,
  Ntotal = 2000,
  BOGs = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{expression_matrix}{A single-cell RNAseq matrix or a list of matrices}

\item{split}{A vector that defines how to split a expression_matrix into multiple ones}

\item{subsample}{A boolean variable that defines if each of the provided
expression_matrixs has to be proportionally subsampled to the same number of cells}

\item{PCs_to_use}{The number of Principal Components to consider for
FindNeighbors of Seurat}

\item{resolution}{The resolution to consider for FindClusters of Seurat}

\item{clusters_mv}{A vector that substitutes the clustering
(ex. celltypes if already annotated)}

\item{Ntotal}{The total number of cells to consider in the subsampling}

\item{BOGs}{A boolean variable to define if map_clusters_to_representation has
to compute also the BOGs}

\item{verbose}{A boolean variable to define if the steps of Seurat will be
printed or not}
}
\value{
The result of map_clusters_to_representation run on each expression_matrix after
splitting, subsampling, clustering if/when required
}
\description{
Signatures Initialization
}
\examples{
set.seed(123)
M <- matrix(runif(200000,0,100), ncol=200)
rownames(M) <- paste0('Gene-', seq(1, dim(M)[1]))
colnames(M) <- paste0('Cell-', seq(1, dim(M)[2]))
getClusterSignatures(M, resolution=2)
}
