% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_eTrace.R
\name{map_eTrace}
\alias{map_eTrace}
\title{Map new profiles to the eTraces}
\usage{
map_eTrace(
  net,
  mapped_obj,
  genes = NULL,
  score = NULL,
  nRand = 100,
  main = NULL,
  n_nearest = 15,
  expression_enrichment = FALSE,
  together = TRUE,
  ylab = "score",
  new_colors = NULL,
  upper_colors = NULL,
  lower_colors = NULL,
  jitter = FALSE,
  factor = NULL
)
}
\arguments{
\item{net}{The reference network}

\item{mapped_obj}{The result of \code{mapNetwork}}

\item{genes}{Genes of which to compute the expression enrichment}

\item{score}{A score to plot on the y axis}

\item{nRand}{The number of random sampling to use (defaults to 100)}

\item{main}{The title of the plot}

\item{n_nearest}{The number of nearest neighbors to consider to derive the x
and y coordinates of the new points}

\item{expression_enrichment}{A boolean, if TRUE the expression enrichment is
computed and plotted (\code{get_eTrace} function). Defaults to FALSE}

\item{together}{A boolean, if FALSE the eTraces plots will be generated for
each mapped point. Defaults to TRUE}

\item{ylab}{The ylab to add. Defaults to \code{score}}

\item{new_colors}{The color(s) of the new points}

\item{upper_colors}{The colors for the upper plot}

\item{lower_colors}{The colors for the lower plot}

\item{jitter}{A boolean, if TRUE (default), it adds a random noise to place
the labels of new points a bit further apart. New runs lead to new positions}

\item{factor}{How wide jitter should be. Defaults to 200}
}
\value{
The plot of the eTraces divided into two subplots, with the new points
highlighted
}
\description{
Map new profiles to the eTraces
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_color <- subclass_palette[net$SubClass]
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_color <- stages_palette[net$Stages]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from, 'to' = edges_to, 'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
SummarizedExperiment::rowData(net)$informative <- sample(c(TRUE, FALSE), size = nrow(net),
replace = TRUE)
new_profiles <- matrix(sample(seq(1,10, length.out=10000), nrow(net)*10, replace = TRUE), ncol = 10)
rownames(new_profiles) <- rownames(net)
colnames(new_profiles) <- paste0('NewCol-', seq(1,10))
mapped_obj <- mapNetwork(net, new_profiles)
map_eTrace(net, mapped_obj)
}
