% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_eTrace.R
\name{map_eTrace}
\alias{map_eTrace}
\title{Map new profiles to the eTraces}
\usage{
map_eTrace(
  net,
  mapped_obj,
  genes = NULL,
  score = NULL,
  nRand = 100,
  main = NULL,
  n_nearest = 15,
  expression_enrichment = FALSE,
  together = TRUE,
  ylab = "score",
  new_colors = NULL,
  upper_colors = NULL,
  lower_colors = NULL
)
}
\arguments{
\item{net}{The reference network}

\item{mapped_obj}{The result of \code{mapNetwork}}

\item{genes}{Genes of which to compute the expression enrichment}

\item{score}{A score to plot on the y axis}

\item{nRand}{The number of random sampling to use (defaults to 100)}

\item{main}{The title of the plot}

\item{n_nearest}{The number of nearest neighbors to consider to derive the x
and y coordinates of the new points}

\item{expression_enrichment}{A boolean, if TRUE the expression enrichment is
computed and plotted (\code{get_eTrace} function). Defaults to FALSE}

\item{together}{A boolean, if FALSE the eTraces plots will be generated for
each mapped point. Defaults to TRUE}

\item{ylab}{The ylab to add. Defaults to \code{score}}

\item{new_colors}{The color(s) of the new points}

\item{upper_colors}{The colors for the upper plot}

\item{lower_colors}{The colors for the lower plot}
}
\value{
The plot of the eTraces divided into two subplots, with the new points
highlighted
}
\description{
Map new profiles to the eTraces
}
