% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eTrace.R
\name{plot_eTrace}
\alias{plot_eTrace}
\title{Plot the eTraces}
\usage{
plot_eTrace(
  net,
  genes = NULL,
  score = NULL,
  expression_enrichment = FALSE,
  main = NULL,
  upper_colors = NULL,
  lower_colors = NULL,
  nRand = 100,
  ylab = "score"
)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the expression enrichment}

\item{score}{A score to plot on the y axis}

\item{expression_enrichment}{A boolean, if TRUE the expression enrichment is
computed and plotted (\code{get_eTrace} function). Defaults to FALSE}

\item{main}{The title of the plot}

\item{upper_colors}{The colors for the upper plot}

\item{lower_colors}{The colors for the lower plot}

\item{nRand}{The number of random sampling to use (defaults to 100)}

\item{ylab}{The ylab to add. Defaults to \code{score}}
}
\value{
The plot of the eTraces divided into two subplots
}
\description{
Plot the eTraces
}
