% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eTrace.R
\name{plot_eTrace}
\alias{plot_eTrace}
\title{Plot the eTraces}
\usage{
plot_eTrace(
  net,
  genes = NULL,
  score = NULL,
  expression_enrichment = FALSE,
  main = NULL,
  upper_colors = NULL,
  lower_colors = NULL,
  nRand = 100,
  ylab = "score"
)
}
\arguments{
\item{net}{The reference network}

\item{genes}{Genes of which to compute the expression enrichment}

\item{score}{A score to plot on the y axis}

\item{expression_enrichment}{A boolean, if TRUE the expression enrichment is
computed and plotted (\code{get_eTrace} function). Defaults to FALSE}

\item{main}{The title of the plot}

\item{upper_colors}{The colors for the upper plot}

\item{lower_colors}{The colors for the lower plot}

\item{nRand}{The number of random sampling to use (defaults to 100)}

\item{ylab}{The ylab to add. Defaults to \code{score}}
}
\value{
The plot of the eTraces divided into two subplots
}
\description{
Plot the eTraces
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_colors <- subclass_palette[net$SubClass]
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_colors <- stages_palette[net$Stages]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from,
'to' = edges_to,
'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
SummarizedExperiment::rowData(net)$informative <- sample(c(TRUE, FALSE), size = nrow(net),
replace = TRUE)
plot_eTrace(net, 'Gene-1')
}
