% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{Create a ggplot of the network}
\usage{
plot_net(
  layout,
  color_attr,
  edges = NULL,
  label_attr = NULL,
  title = NULL,
  new_points_col = "#FF0000",
  legend = FALSE,
  legend_name = NULL,
  only_new_points = FALSE,
  alpha = NULL,
  show_edges = TRUE,
  n_increase = 1,
  stroke = 0.2
)
}
\arguments{
\item{layout}{A data frame with the layout informations to plot}

\item{color_attr}{The color attribute}

\item{edges}{The edges to visualize}

\item{label_attr}{A label vector for the clusters}

\item{title}{A title}

\item{new_points_col}{The color of newly mapped points}

\item{legend}{A boolean, if TRUE the legend is shown. Defaults to FALSE}

\item{legend_name}{The legend name}

\item{only_new_points}{A boolean, if TRUE only new points are labeled. Defaults
to FALSE}

\item{alpha}{A vector with the alpha information. Defaults to all 1}

\item{show_edges}{A boolean, if FALSE no edges are shown}

\item{n_increase}{The increase for newly mapped points. Defaults to 1}

\item{stroke}{The stroke around the points, defaults to 0.2}
}
\value{
A ggplot
}
\description{
Create a ggplot of the network
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_colors <- subclass_palette[net$SubClass]
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_colors <- stages_palette[net$Stages]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from,
'to' = edges_to,
'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
SummarizedExperiment::rowData(net)$informative <- sample(c(TRUE, FALSE), size = nrow(net),
replace = TRUE)
l <- cbind(net$X_coord, net$Y_coord, 'Size' = rep(3, dim(net)[1]),
'Color' = net$SubClass_colors, 'Group' = net$SubClass)
neuRoDev:::plot_net(l, net$SubClass_colors, edges = edges_df)
}
