% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkScore.R
\name{plotNetworkScore}
\alias{plotNetworkScore}
\title{Plot a score onto the network}
\usage{
plotNetworkScore(
  net,
  genes = NULL,
  score = NULL,
  label_attr = NULL,
  smooth = FALSE,
  n_nearest = 15,
  normalize = TRUE,
  na.vec = NULL,
  palette = grDevices::blues9,
  title = NULL,
  show_edges = TRUE,
  stroke = 0.5,
  fix_alpha = FALSE,
  max_size = 3
)
}
\arguments{
\item{net}{The reference network}

\item{genes}{The genes to consider for the score. If multiple genes are given,
the score will be their average.}

\item{score}{The score to plot. If NULL (default), the logcounts of the
pseudobulks are used}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}

\item{smooth}{A boolean, if TRUE (default), the scores are smoothed to consider
the network structure}

\item{n_nearest}{The number of nearest neighbors to consider for smoothing.
Defaults to 15}

\item{normalize}{A boolean, if TRUE (default) and score is NULL, it normalizes
the pseudobulk expression computing a row-wise z-score}

\item{na.vec}{The vector to define if certain clusters are not to be shown}

\item{palette}{The color palette to use. Defaults to \code{blues9}}

\item{title}{The title of the plot}

\item{show_edges}{A boolean, if FALSE no edges are shown. Defaults to TRUE}

\item{stroke}{The stroke of the points, defaults to 0.5}

\item{fix_alpha}{A boolean, if TRUE all points will have alpha 1, otherwise
it will be proportional to the score. Defaults to FALSE}

\item{max_size}{The maximum size of the dots in the network visualization}
}
\value{
A ggplot of the network colored based on the score
}
\description{
Plot a score onto the network
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_colors <- subclass_palette[net$SubClass]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from, 'to' = edges_to, 'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
plotNetworkScore(net, 'Gene-1')
}
