% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSameLayout.R
\name{plotSameLayout}
\alias{plotSameLayout}
\title{Plot a mapped dataset conserving the same layout as the original reference
network}
\usage{
plotSameLayout(
  net,
  new_cor,
  color_attr = "SubClass",
  label_attr = "SubClass",
  col_vector = NULL,
  no_label = FALSE,
  new_points_col = "#FF0000",
  max_size = 3,
  only_new_points = FALSE,
  mapping_quality = TRUE,
  annotate = TRUE,
  n_nearest = 15,
  new_name = NULL,
  order_names = FALSE,
  ...
)
}
\arguments{
\item{net}{The reference network}

\item{new_cor}{The correlation matrix between the profiles and the reference
pseudobulks}

\item{color_attr}{The color attribute in \code{net}, defaults to \code{SubClass}}

\item{label_attr}{The label attribute in \code{net}, defaults to \code{SubClass}}

\item{col_vector}{The color vector in \code{net}, defaults to \code{SubClass_color}}

\item{no_label}{A boolean, if TRUE the clusters' labels will not be shown.
Defaults to FALSE}

\item{new_points_col}{The color of the mapped points}

\item{max_size}{The maximum size of the dots in the network visualization}

\item{only_new_points}{A boolean, if TRUE only the labels of the new points
are shown. Defaults to FALSE}

\item{mapping_quality}{A boolean, if TRUE (default), it returns also the
mapping quality assessment}

\item{annotate}{A boolean, if TRUE (default), it returns also the
annotation scores for the mapped points}

\item{n_nearest}{The number of nearest neighbors to consider for the
layout definition, mapping quality assessment and annotation. Defaults to 15}

\item{new_name}{A vector to define the names of the new profiles. Defaults to
their colnames}

\item{order_names}{A boolean, if TRUE the names in annotateMapping
barplot are ordered. Defaults to FALSE}

\item{...}{Additional parameters for \code{plot_net}}
}
\value{
An \code{S4Vectors} list containing the correlation matrix between the
profiles and the reference pseudobulks (\code{new_cor}), a data.frame containing
the edges information to connect the new profiles with the pseudobulks in the
network (\code{new_edges}), a matrix containing the layout information, including
the x and y coordinates (\code{new_layout}), the new plot (\code{new_plot}), a mapping
quality assessment (if mapping_quality = TRUE) (\code{mapping_quality}), and
the annotation scores (if annotate = TRUE) (\code{annotation})
}
\description{
Plot a mapped dataset conserving the same layout as the original reference
network
}
\examples{
m <- matrix(sample(seq(1,10, length.out=10000), 15000*100, replace = TRUE), ncol = 100)
rownames(m) <- paste0('Gene-', seq(1,15000))
colnames(m) <- paste0('Col-', seq(1,100))
net <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = m))
net$SubClass <- rep(c('A', 'B', 'C', 'D'), each = 25)
subclass_palette <- c('A' = 'red', 'B' = 'blue', 'C' = 'green', 'D' = 'yellow')
net$SubClass_color <- subclass_palette[net$SubClass]
net$Stages <- rep(c('S1', 'S2', 'S3', 'S4'), each = 25)
stages_palette <- c('S1' = 'pink', 'S2' = 'orange', 'S3' = 'violet', 'S4' = 'black')
net$Stages_color <- stages_palette[net$Stages]
net$X_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
net$Y_coord <- sample(seq(1,2, length.out = 1000), size = ncol(net), replace = TRUE)
edges_from <- sample(colnames(net), size = 200, replace = TRUE)
edges_to <- sample(colnames(net), size = 200, replace = TRUE)
edges_from_x <- net$X_coord[match(edges_from, colnames(net))]
edges_from_y <- net$Y_coord[match(edges_from, colnames(net))]
edges_to_x <- net$X_coord[match(edges_to, colnames(net))]
edges_to_y <- net$Y_coord[match(edges_to, colnames(net))]
edges_weight <- sample(seq(0,1, length.out=1000), length(edges_from), replace = TRUE)
edges_df <- data.frame('from' = edges_from,
'to' = edges_to,
'weight' = edges_weight,
'from.x' = edges_from_x,
'from.y' = edges_from_y,
'to.x' = edges_to_x,
'to.y' = edges_to_y)
net@metadata$network$edges <- edges_df
SummarizedExperiment::rowData(net)$informative <- sample(c(TRUE, FALSE), size = nrow(net),
replace = TRUE)
new_profiles <- matrix(sample(seq(1,10, length.out=10000), nrow(net)*10, replace = TRUE), ncol = 10)
rownames(new_profiles) <- rownames(net)
colnames(new_profiles) <- paste0('NewCol-', seq(1,10))
common_genes <- intersect(rownames(net)[SingleCellExperiment::rowData(net)$informative],
rownames(new_profiles))
new_cor <- stats::cor(t(apply(as.matrix(SingleCellExperiment::logcounts(net)[common_genes,]),
1, function(v) {(v-mean(v))/stats::sd(v)})),new_profiles[common_genes,])
neuRoDev:::plotSameLayout(net, new_cor)
}
